refactor(storage,query): split oversized repo.rs and engine.rs modules (#67)
Pure mechanical move, no behavior change. Public APIs unchanged.

storage/src/repo.rs (2121 lines) -> storage/src/repo/
  mod.rs (726), replay.rs (191), backup.rs (204),
  search.rs (106), transaction.rs (193), tests.rs (734)

query/src/engine.rs (1837 lines) -> query/src/engine/
  mod.rs (437), planning.rs (368), execution.rs (718), synthesis.rs (320)

All files now under the 800-line guideline.
Also replaced rkyv::Infallible .unwrap() with .expect("infallible deserializer")
in the WAL replay/backup path to document intent.

Verified: cargo fmt --check, clippy -D warnings clean, 241 workspace tests pass.
