--- query/src/engine.rs
@@ -1,1837 +1,0 @@
-use crate::dsl::{QueryMode, QueryRequest, SearchMode};
-use crate::graphrag::{
-    compute_groundedness, map_community_summaries, reduce_community_summaries, GroundednessInput,
-    DRIFT_EVIDENCE_THRESHOLD, DRIFT_MAX_ITERATIONS,
-};
-use crate::planner::{QueryPlan, QueryPlanner};
-use crate::semantic_cache::{SemanticCache, SemanticCacheConfig, SemanticCacheKey};
-use alayasiki_core::audit::{AuditEvent, AuditOperation, AuditOutcome, AuditSink};
-use alayasiki_core::auth::{
-    Action, AuthError, Authorizer, AuthzError, JwtAuthenticator, Principal, ResourceContext,
-};
-use alayasiki_core::embedding::deterministic_embedding;
-use alayasiki_core::error::{AlayasikiError, ErrorCode};
-use alayasiki_core::metrics::{MetricsCollector, MetricsSnapshot};
-use alayasiki_core::model::Node;
-use chrono::{DateTime, NaiveDate, Utc};
-use serde::{Deserialize, Serialize};
-use std::cmp::Ordering;
-use std::collections::{BTreeSet, HashMap, HashSet, VecDeque};
-use std::sync::Arc;
-use std::time::Instant;
-use storage::community::CommunitySummary;
-use storage::repo::{parse_wal_snapshot_lsn, RepoError, Repository, SnapshotView};
-use storage::session::{SessionGraph, SessionOwner};
-use thiserror::Error;
-use tokio::sync::Mutex;
-
-/// Provenance metadata attached to evidence items.
-/// Captures the data lineage: where it came from and how it was extracted.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize, Default)]
-pub struct Provenance {
-    /// Original source document (e.g. "s3://bucket/file.pdf")
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub source: Option<String>,
-    /// Model that extracted/processed this data
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub extraction_model_id: Option<String>,
-    /// Snapshot when this data was ingested
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub snapshot_id: Option<String>,
-    /// Timestamp of ingestion (RFC3339)
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub ingested_at: Option<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct EvidenceNode {
-    pub id: u64,
-    pub data: String,
-    pub score: f32,
-    pub hop: u8,
-    pub provenance: Provenance,
-    pub confidence: f32,
-}
-
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct EvidenceEdge {
-    pub source: u64,
-    pub target: u64,
-    pub relation: String,
-    pub weight: f32,
-    pub provenance: Provenance,
-    pub confidence: f32,
-}
-
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct EvidenceSubgraph {
-    pub nodes: Vec<EvidenceNode>,
-    pub edges: Vec<EvidenceEdge>,
-}
-
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct Citation {
-    pub source: String,
-    pub span: [usize; 2],
-    pub node_id: u64,
-    pub confidence: f32,
-}
-
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct Anchor {
-    pub node_id: u64,
-    pub score: f32,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct ExpansionPath {
-    pub anchor_id: u64,
-    pub target_id: u64,
-    pub path: Vec<u64>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct ExclusionReason {
-    pub node_id: Option<u64>,
-    pub reason: String,
-}
-
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct ExplainPlan {
-    pub steps: Vec<String>,
-    pub effective_search_mode: SearchMode,
-    pub anchors: Vec<Anchor>,
-    pub expansion_paths: Vec<ExpansionPath>,
-    pub exclusions: Vec<ExclusionReason>,
-}
-
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct QueryResponse {
-    pub answer: Option<String>,
-    pub evidence: EvidenceSubgraph,
-    pub citations: Vec<Citation>,
-    pub groundedness: f32,
-    pub explain: ExplainPlan,
-    pub model_id: Option<String>,
-    pub snapshot_id: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub time_travel: Option<String>,
-    pub latency_ms: u64,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub error_code: Option<ErrorCode>,
-}
-
-#[derive(Debug, Error)]
-pub enum QueryError {
-    #[error("invalid query: {0}")]
-    InvalidQuery(String),
-    #[error("not found: {0}")]
-    NotFound(String),
-    #[error("repository error: {0}")]
-    Repository(#[from] RepoError),
-    #[error("authorization error: {0}")]
-    Unauthorized(#[from] AuthzError),
-    #[error("authentication error: {0}")]
-    Unauthenticated(#[from] AuthError),
-}
-
-impl AlayasikiError for QueryError {
-    fn error_code(&self) -> ErrorCode {
-        match self {
-            QueryError::InvalidQuery(_) => ErrorCode::InvalidArgument,
-            QueryError::NotFound(_) => ErrorCode::NotFound,
-            QueryError::Repository(err) => err.error_code(),
-            QueryError::Unauthorized(err) => err.error_code(),
-            QueryError::Unauthenticated(err) => err.error_code(),
-        }
-    }
-}
-
-impl QueryError {
-    pub fn to_response(&self) -> QueryResponse {
-        QueryResponse {
-            answer: Some(self.to_string()),
-            evidence: EvidenceSubgraph {
-                nodes: vec![],
-                edges: vec![],
-            },
-            citations: vec![],
-            groundedness: 0.0,
-            explain: ExplainPlan {
-                steps: vec!["error".to_string()],
-                effective_search_mode: SearchMode::Auto,
-                anchors: vec![],
-                expansion_paths: vec![],
-                exclusions: vec![],
-            },
-            model_id: None,
-            snapshot_id: None,
-            time_travel: None,
-            latency_ms: 0,
-            error_code: Some(self.error_code()),
-        }
-    }
-}
-
-pub struct QueryEngine {
-    repo: Arc<Repository>,
-    community_summaries: Vec<CommunitySummary>,
-    audit_sink: Option<Arc<dyn AuditSink>>,
-    semantic_cache: Arc<Mutex<SemanticCache<QueryResponse>>>,
-    metrics: Arc<MetricsCollector>,
-}
-
-const DEFAULT_EMBEDDING_MODEL_ID: &str = "embedding-default-v1";
-const UNICODE_NGRAM_SIZE: usize = 2;
-
-#[derive(Debug, Clone)]
-pub struct RankedNode {
-    pub id: u64,
-    pub data: String,
-    pub score: f32,
-    pub hop: u8,
-    pub source: Option<String>,
-    pub extraction_model_id: Option<String>,
-    pub node_snapshot_id: Option<String>,
-    pub ingested_at: Option<String>,
-    pub confidence: f32,
-}
-
-/// Internal edge representation during query execution (before final output).
-#[derive(Debug, Clone)]
-pub struct InternalEdge {
-    pub source: u64,
-    pub target: u64,
-    pub relation: String,
-    pub weight: f32,
-    pub provenance: Provenance,
-    pub confidence: f32,
-}
-
-#[derive(Debug, Clone)]
-pub struct ExecutionState {
-    pub anchors: Vec<Anchor>,
-    pub expansion_paths: Vec<ExpansionPath>,
-    pub exclusions: Vec<ExclusionReason>,
-    pub nodes: Vec<RankedNode>,
-    pub edges: Vec<InternalEdge>,
-}
-
-#[derive(Clone)]
-struct ResolvedSnapshot {
-    snapshot_id: String,
-    snapshot_lsn: u64,
-    snapshot_view: Option<Arc<SnapshotView>>,
-    time_travel: Option<String>,
-    requires_versioned_summaries: bool,
-}
-
-impl QueryEngine {
-    pub fn new(repo: Arc<Repository>) -> Self {
-        Self {
-            repo,
-            community_summaries: Vec::new(),
-            audit_sink: None,
-            semantic_cache: Arc::new(Mutex::new(SemanticCache::with_config(
-                SemanticCacheConfig::default(),
-            ))),
-            metrics: Arc::new(MetricsCollector::new(1000)),
-        }
-    }
-
-    /// Attach pre-computed community summaries for global search support.
-    pub fn with_community_summaries(mut self, summaries: Vec<CommunitySummary>) -> Self {
-        self.community_summaries = summaries;
-        self
-    }
-
-    pub fn with_audit_sink(mut self, sink: Arc<dyn AuditSink>) -> Self {
-        self.audit_sink = Some(sink);
-        self
-    }
-
-    pub fn with_semantic_cache_config(mut self, config: SemanticCacheConfig) -> Self {
-        self.semantic_cache = Arc::new(Mutex::new(SemanticCache::with_config(config)));
-        self
-    }
-
-    pub fn metrics(&self) -> MetricsSnapshot {
-        self.metrics.snapshot()
-    }
-
-    pub fn metrics_collector(&self) -> Arc<MetricsCollector> {
-        self.metrics.clone()
-    }
-
-    pub async fn execute_json(&self, raw: &str) -> Result<QueryResponse, QueryError> {
-        let request = QueryRequest::parse_json(raw)
-            .map_err(|err| QueryError::InvalidQuery(err.to_string()))?;
-        self.execute(request).await
-    }
-
-    pub async fn execute_json_authorized(
-        &self,
-        raw: &str,
-        principal: &Principal,
-        authorizer: &Authorizer,
-        resource: &ResourceContext,
-    ) -> Result<QueryResponse, QueryError> {
-        let request = QueryRequest::parse_json(raw)
-            .map_err(|err| QueryError::InvalidQuery(err.to_string()))?;
-        self.execute_authorized(request, principal, authorizer, resource)
-            .await
-    }
-
-    pub async fn execute_json_jwt_authorized(
-        &self,
-        raw: &str,
-        bearer_token: &str,
-        authenticator: &JwtAuthenticator,
-        authorizer: &Authorizer,
-        resource: &ResourceContext,
-    ) -> Result<QueryResponse, QueryError> {
-        let principal = self.authenticate_query_principal(
-            bearer_token,
-            authenticator,
-            DEFAULT_EMBEDDING_MODEL_ID,
-        )?;
-        let request = QueryRequest::parse_json(raw)
-            .map_err(|err| QueryError::InvalidQuery(err.to_string()))?;
-        self.execute_authorized(request, &principal, authorizer, resource)
-            .await
-    }
-
-    pub async fn execute_authorized(
-        &self,
-        request: QueryRequest,
-        principal: &Principal,
-        authorizer: &Authorizer,
-        resource: &ResourceContext,
-    ) -> Result<QueryResponse, QueryError> {
-        let model_id = effective_query_model_id(&request);
-        if let Err(err) = authorizer.authorize(principal, Action::Query, resource) {
-            self.emit_audit_event(build_query_audit_event(
-                AuditOutcome::Denied,
-                &model_id,
-                Some(principal.subject.clone()),
-                Some(principal.tenant.clone()),
-                None,
-                Some(err.to_string()),
-            ));
-            return Err(err.into());
-        }
-
-        self.execute_with_audit(
-            request,
-            Some(principal.subject.clone()),
-            Some(principal.tenant.clone()),
-            Some(principal.tenant.clone()),
-            Some(SessionOwner::new(
-                principal.tenant.clone(),
-                principal.subject.clone(),
-            )),
-        )
-        .await
-    }
-
-    pub async fn execute_jwt_authorized(
-        &self,
-        request: QueryRequest,
-        bearer_token: &str,
-        authenticator: &JwtAuthenticator,
-        authorizer: &Authorizer,
-        resource: &ResourceContext,
-    ) -> Result<QueryResponse, QueryError> {
-        let model_id = effective_query_model_id(&request);
-        let principal =
-            self.authenticate_query_principal(bearer_token, authenticator, &model_id)?;
-
-        self.execute_authorized(request, &principal, authorizer, resource)
-            .await
-    }
-
-    fn authenticate_query_principal(
-        &self,
-        bearer_token: &str,
-        authenticator: &JwtAuthenticator,
-        model_id: &str,
-    ) -> Result<Principal, QueryError> {
-        authenticator.authenticate(bearer_token).map_err(|err| {
-            self.emit_audit_event(build_query_audit_event(
-                AuditOutcome::Denied,
-                model_id,
-                None,
-                None,
-                None,
-                Some(err.to_string()),
-            ));
-            err.into()
-        })
-    }
-
-    pub async fn execute(&self, request: QueryRequest) -> Result<QueryResponse, QueryError> {
-        self.execute_with_audit(request, None, None, None, None)
-            .await
-    }
-
-    async fn execute_with_audit(
-        &self,
-        request: QueryRequest,
-        actor: Option<String>,
-        tenant: Option<String>,
-        tenant_scope: Option<String>,
-        session_owner: Option<SessionOwner>,
-    ) -> Result<QueryResponse, QueryError> {
-        let start = Instant::now();
-        let model_id = effective_query_model_id(&request);
-        let result = self
-            .execute_internal(request, start, tenant_scope, session_owner)
-            .await;
-        match &result {
-            Ok(response) => {
-                self.emit_audit_event(build_query_audit_event(
-                    AuditOutcome::Succeeded,
-                    &model_id,
-                    actor,
-                    tenant,
-                    response.snapshot_id.clone(),
-                    None,
-                ));
-            }
-            Err(err) => {
-                self.emit_audit_event(build_query_audit_event(
-                    AuditOutcome::Failed,
-                    &model_id,
-                    actor,
-                    tenant,
-                    None,
-                    Some(err.to_string()),
-                ));
-                self.metrics
-                    .record_query(start.elapsed().as_micros() as u64, false);
-            }
-        }
-        result
-    }
-
-    async fn execute_internal(
-        &self,
-        request: QueryRequest,
-        start: Instant,
-        tenant_scope: Option<String>,
-        session_owner: Option<SessionOwner>,
-    ) -> Result<QueryResponse, QueryError> {
-        request
-            .validate()
-            .map_err(|err| QueryError::InvalidQuery(err.to_string()))?;
-
-        let effective_model_id = request
-            .model_id
-            .clone()
-            .unwrap_or_else(|| DEFAULT_EMBEDDING_MODEL_ID.to_string());
-        let mut plan = QueryPlanner::plan(&request);
-        let resolved_snapshot = self.resolve_snapshot(&request).await?;
-        let tenant_scoped = tenant_scope.is_some();
-        let cache_eligible = !tenant_scoped && request.session_id.is_none();
-
-        // Resolve session if provided.
-        let session_graph = match request.session_id.as_deref() {
-            Some(session_id) => self
-                .repo
-                .get_session_with_owner(session_id, session_owner.as_ref())?,
-            None => None,
-        };
-        let cache_key = SemanticCacheKey::from_request(
-            &request,
-            &effective_model_id,
-            &resolved_snapshot.snapshot_id,
-            plan.effective_search_mode,
-        );
-
-        if cache_eligible {
-            if let Some(mut cached_response) =
-                self.lookup_semantic_cache(&cache_key, &request.query).await
-            {
-                cached_response.latency_ms = start.elapsed().as_millis() as u64;
-                if !cached_response
-                    .explain
-                    .steps
-                    .iter()
-                    .any(|step| step == crate::SEMANTIC_CACHE_HIT_STEP)
-                {
-                    cached_response
-                        .explain
-                        .steps
-                        .insert(0, crate::SEMANTIC_CACHE_HIT_STEP.to_string());
-                }
-
-                self.metrics
-                    .record_query(start.elapsed().as_micros() as u64, true);
-                return Ok(cached_response);
-            }
-        }
-
-        // Dispatch to the appropriate search mode pipeline.
-        let (state, plan, global_answer) = match plan.effective_search_mode {
-            SearchMode::Global => {
-                self.execute_global(
-                    &request,
-                    &mut plan,
-                    &effective_model_id,
-                    &resolved_snapshot,
-                    tenant_scope.as_deref(),
-                    session_graph.as_ref(),
-                )
-                .await?
-            }
-            SearchMode::Drift => {
-                let (state, plan) = self
-                    .execute_drift(
-                        &request,
-                        &mut plan,
-                        &effective_model_id,
-                        resolved_snapshot.snapshot_view.as_deref(),
-                        tenant_scope.as_deref(),
-                        session_graph.as_ref(),
-                    )
-                    .await?;
-                (state, plan, None)
-            }
-            SearchMode::Local | SearchMode::Auto => {
-                let (state, plan) = self
-                    .execute_local_with_auto_fallback(
-                        &request,
-                        plan,
-                        &effective_model_id,
-                        resolved_snapshot.snapshot_view.as_deref(),
-                        tenant_scope.as_deref(),
-                        session_graph.as_ref(),
-                    )
-                    .await?;
-                (state, plan, None)
-            }
-        };
-
-        // Build evidence nodes with provenance from execution state.
-        let evidence_nodes: Vec<EvidenceNode> = state
-            .nodes
-            .iter()
-            .map(|node| EvidenceNode {
-                id: node.id,
-                data: node.data.clone(),
-                score: node.score,
-                hop: node.hop,
-                provenance: Provenance {
-                    source: node.source.clone(),
-                    extraction_model_id: node.extraction_model_id.clone(),
-                    snapshot_id: node.node_snapshot_id.clone(),
-                    ingested_at: node.ingested_at.clone(),
-                },
-                confidence: node.confidence,
-            })
-            .collect();
-
-        // Build evidence edges with provenance.
-        let evidence_edges: Vec<EvidenceEdge> = state
-            .edges
-            .iter()
-            .map(|edge| {
-                // Look up edge metadata from the graph index via repo nodes
-                // Edge confidence defaults to weight
-                EvidenceEdge {
-                    source: edge.source,
-                    target: edge.target,
-                    relation: edge.relation.clone(),
-                    weight: edge.weight,
-                    provenance: edge.provenance.clone(),
-                    confidence: edge.confidence,
-                }
-            })
-            .collect();
-
-        let citations = build_citations(&state.nodes);
-
-        // Improved groundedness scoring.
-        let evidence_scores: Vec<f32> = state.nodes.iter().map(|n| n.score).collect();
-        let source_diversity = {
-            let sources: HashSet<&str> = state
-                .nodes
-                .iter()
-                .filter_map(|n| n.source.as_deref())
-                .collect();
-            sources.len()
-        };
-        let has_graph_support = !evidence_edges.is_empty();
-        let groundedness = compute_groundedness(&GroundednessInput {
-            query: &request.query,
-            evidence_scores: &evidence_scores,
-            evidence_count: evidence_nodes.len(),
-            source_diversity,
-            has_graph_support,
-        });
-
-        // Generate answer based on mode.
-        let answer = match request.mode {
-            QueryMode::Evidence => None,
-            QueryMode::Answer => {
-                if let Some(global_ans) = global_answer {
-                    Some(global_ans)
-                } else {
-                    Some(generate_answer(&request.query, &evidence_nodes))
-                }
-            }
-        };
-
-        let latency_ms = start.elapsed().as_millis() as u64;
-
-        let response = QueryResponse {
-            answer,
-            evidence: EvidenceSubgraph {
-                nodes: evidence_nodes,
-                edges: evidence_edges,
-            },
-            citations,
-            groundedness,
-            explain: ExplainPlan {
-                steps: plan.steps.iter().map(|step| step.to_string()).collect(),
-                effective_search_mode: plan.effective_search_mode,
-                anchors: state.anchors,
-                expansion_paths: state.expansion_paths,
-                exclusions: state.exclusions,
-            },
-            model_id: Some(effective_model_id),
-            snapshot_id: Some(resolved_snapshot.snapshot_id.clone()),
-            time_travel: resolved_snapshot.time_travel.clone(),
-            latency_ms,
-            error_code: None,
-        };
-
-        self.metrics.record_query(
-            start.elapsed().as_micros() as u64,
-            response
-                .explain
-                .steps
-                .iter()
-                .any(|s| s == crate::SEMANTIC_CACHE_HIT_STEP),
-        );
-
-        if cache_eligible {
-            self.insert_semantic_cache(cache_key, &request.query, response.clone())
-                .await;
-        }
-
-        Ok(response)
-    }
-
-    fn emit_audit_event(&self, event: AuditEvent) {
-        if let Some(sink) = &self.audit_sink {
-            let _ = sink.record(event);
-        }
-    }
-
-    async fn lookup_semantic_cache(
-        &self,
-        key: &SemanticCacheKey,
-        query: &str,
-    ) -> Option<QueryResponse> {
-        let mut cache = self.semantic_cache.lock().await;
-        cache.lookup(key, query)
-    }
-
-    async fn insert_semantic_cache(
-        &self,
-        key: SemanticCacheKey,
-        query: &str,
-        response: QueryResponse,
-    ) {
-        let mut cache = self.semantic_cache.lock().await;
-        cache.insert(key, query, response);
-    }
-
-    async fn resolve_snapshot(
-        &self,
-        request: &QueryRequest,
-    ) -> Result<ResolvedSnapshot, QueryError> {
-        if let Some(snapshot_id) = request.snapshot_id.clone() {
-            let snapshot_lsn = parse_wal_snapshot_lsn(&snapshot_id).ok_or_else(|| {
-                QueryError::InvalidQuery(format!(
-                    "snapshot_id must be wal-lsn-<lsn>: {snapshot_id}"
-                ))
-            })?;
-            let snapshot_view = self.load_snapshot_view(&snapshot_id).await?;
-            return Ok(ResolvedSnapshot {
-                snapshot_id,
-                snapshot_lsn,
-                snapshot_view: Some(snapshot_view),
-                time_travel: None,
-                requires_versioned_summaries: true,
-            });
-        }
-
-        if let Some(time_travel) = request.time_travel.as_deref() {
-            let as_of_unix_ms = parse_time_travel_as_of_unix_ms(time_travel)?;
-            let snapshot_id = self
-                .repo
-                .resolve_snapshot_id_at_or_before(as_of_unix_ms)
-                .await
-                .map_err(|err| match err {
-                    RepoError::SnapshotNotFound(_) => {
-                        QueryError::NotFound(format!("time_travel `{time_travel}`"))
-                    }
-                    other => QueryError::Repository(other),
-                })?;
-            let snapshot_lsn = parse_wal_snapshot_lsn(&snapshot_id).ok_or_else(|| {
-                QueryError::InvalidQuery(format!(
-                    "snapshot_id must be wal-lsn-<lsn>: {snapshot_id}"
-                ))
-            })?;
-            let snapshot_view = self.load_snapshot_view(&snapshot_id).await?;
-            return Ok(ResolvedSnapshot {
-                snapshot_id,
-                snapshot_lsn,
-                snapshot_view: Some(snapshot_view),
-                time_travel: Some(time_travel.to_string()),
-                requires_versioned_summaries: true,
-            });
-        }
-
-        let snapshot_id = self.repo.current_snapshot_id().await;
-        let snapshot_lsn = parse_wal_snapshot_lsn(&snapshot_id).ok_or_else(|| {
-            QueryError::InvalidQuery(format!("snapshot_id must be wal-lsn-<lsn>: {snapshot_id}"))
-        })?;
-        Ok(ResolvedSnapshot {
-            snapshot_id,
-            snapshot_lsn,
-            snapshot_view: None,
-            time_travel: None,
-            requires_versioned_summaries: false,
-        })
-    }
-
-    async fn load_snapshot_view(&self, snapshot_id: &str) -> Result<Arc<SnapshotView>, QueryError> {
-        match self.repo.load_snapshot_view(snapshot_id).await {
-            Ok(view) => Ok(Arc::new(view)),
-            Err(RepoError::SnapshotNotFound(_)) => {
-                Err(QueryError::NotFound(format!("snapshot_id `{snapshot_id}`")))
-            }
-            Err(RepoError::InvalidSnapshotId(_)) => Err(QueryError::InvalidQuery(format!(
-                "snapshot_id must be wal-lsn-<lsn>: {snapshot_id}"
-            ))),
-            Err(err) => Err(QueryError::Repository(err)),
-        }
-    }
-
-    async fn list_node_ids_from_source(
-        &self,
-        snapshot_view: Option<&SnapshotView>,
-        session: Option<&SessionGraph>,
-    ) -> Vec<u64> {
-        let mut out = match snapshot_view {
-            Some(view) => view.list_node_ids(),
-            None => self.repo.list_node_ids().await,
-        };
-        if let Some(session) = session {
-            out.extend(session.nodes.keys().copied());
-            out.sort_unstable();
-            out.dedup();
-        }
-        out
-    }
-
-    async fn get_nodes_by_ids_from_source(
-        &self,
-        ids: &[u64],
-        snapshot_view: Option<&SnapshotView>,
-        session: Option<&SessionGraph>,
-    ) -> Vec<Node> {
-        let mut results = Vec::with_capacity(ids.len());
-        let mut remaining_ids = Vec::new();
-
-        if let Some(session) = session {
-            for id in ids {
-                if let Some(node) = session.nodes.get(id) {
-                    results.push(node.clone());
-                } else {
-                    remaining_ids.push(*id);
-                }
-            }
-        } else {
-            remaining_ids = ids.to_vec();
-        }
-
-        if !remaining_ids.is_empty() {
-            let mut source_results = match snapshot_view {
-                Some(view) => view.get_nodes_by_ids(&remaining_ids),
-                None => self.repo.get_nodes_by_ids(&remaining_ids).await,
-            };
-            results.append(&mut source_results);
-        }
-        results
-    }
-
-    async fn get_edge_metadata_bulk_from_source(
-        &self,
-        keys: &[(u64, u64, String)],
-        snapshot_view: Option<&SnapshotView>,
-    ) -> HashMap<(u64, u64, String), HashMap<String, String>> {
-        match snapshot_view {
-            Some(view) => view.get_edge_metadata_bulk(keys),
-            None => self.repo.get_edge_metadata_bulk(keys).await,
-        }
-    }
-
-    // -----------------------------------------------------------------------
-    // Local Search with Auto-fallback to DRIFT
-    // -----------------------------------------------------------------------
-    async fn execute_local_with_auto_fallback(
-        &self,
-        request: &QueryRequest,
-        mut plan: QueryPlan,
-        embedding_model_id: &str,
-        snapshot_view: Option<&SnapshotView>,
-        tenant_scope: Option<&str>,
-        session: Option<&SessionGraph>,
-    ) -> Result<(ExecutionState, QueryPlan), QueryError> {
-        let mut state = self
-            .execute_with_plan(
-                request,
-                &plan,
-                embedding_model_id,
-                snapshot_view,
-                tenant_scope,
-                session,
-            )
-            .await?;
-
-        // Record vector-only fallback when graph expansion added nothing.
-        if state.edges.is_empty() && !state.nodes.is_empty() {
-            state.exclusions.push(ExclusionReason {
-                node_id: None,
-                reason: "no_graph_expansion_vector_only_fallback".to_string(),
-            });
-        }
-
-        // Auto mode: fall back to DRIFT if local yields insufficient evidence.
-        if request.search_mode == SearchMode::Auto
-            && plan.effective_search_mode == SearchMode::Local
-            && state.nodes.len() < 2
-        {
-            let (drift_state, drift_plan) = self
-                .execute_drift(
-                    request,
-                    &mut plan,
-                    embedding_model_id,
-                    snapshot_view,
-                    tenant_scope,
-                    session,
-                )
-                .await?;
-            let mut drift_state = drift_state;
-            drift_state.exclusions.push(ExclusionReason {
-                node_id: None,
-                reason: "auto_fallback_to_drift_due_to_insufficient_evidence".to_string(),
-            });
-
-            return Ok((drift_state, drift_plan));
-        }
-
-        Ok((state, plan))
-    }
-
-    // -----------------------------------------------------------------------
-    // Global Search: Community Summary Map-Reduce
-    // -----------------------------------------------------------------------
-    async fn execute_global(
-        &self,
-        request: &QueryRequest,
-        plan: &mut QueryPlan,
-        embedding_model_id: &str,
-        resolved_snapshot: &ResolvedSnapshot,
-        tenant_scope: Option<&str>,
-        session: Option<&SessionGraph>,
-    ) -> Result<(ExecutionState, QueryPlan, Option<String>), QueryError> {
-        let snapshot_view = resolved_snapshot.snapshot_view.as_deref();
-        if tenant_scope.is_some() {
-            // Community summaries are currently shared across tenants.
-            // Disable summary synthesis under tenant-scoped authorization to avoid leakage.
-            plan.steps = vec![
-                "vector_search",
-                "graph_expansion",
-                "context_pruning",
-                "global_fallback_tenant_scoped",
-            ];
-            let mut state = self
-                .execute_with_plan(
-                    request,
-                    plan,
-                    embedding_model_id,
-                    snapshot_view,
-                    tenant_scope,
-                    session,
-                )
-                .await?;
-            state.exclusions.push(ExclusionReason {
-                node_id: None,
-                reason: "global_summary_disabled_by_tenant_scope".to_string(),
-            });
-            return Ok((state, plan.clone(), None));
-        }
-
-        if self.community_summaries.is_empty() {
-            // Fallback: no community data available — run expanded vector search.
-            plan.steps = vec![
-                "vector_search",
-                "graph_expansion",
-                "context_pruning",
-                "global_fallback_no_community_data",
-            ];
-            let mut state = self
-                .execute_with_plan(
-                    request,
-                    plan,
-                    embedding_model_id,
-                    snapshot_view,
-                    tenant_scope,
-                    session,
-                )
-                .await?;
-            state.exclusions.push(ExclusionReason {
-                node_id: None,
-                reason: "no_community_data_fallback_to_vector".to_string(),
-            });
-            return Ok((state, plan.clone(), None));
-        }
-
-        let summary_candidates: Vec<CommunitySummary> = self
-            .community_summaries
-            .iter()
-            .filter(|summary| {
-                summary.is_visible_at_lsn(resolved_snapshot.snapshot_lsn)
-                    && (!resolved_snapshot.requires_versioned_summaries
-                        || summary.snapshot_lsn_range.is_some())
-            })
-            .cloned()
-            .collect();
-        if summary_candidates.is_empty() && resolved_snapshot.requires_versioned_summaries {
-            plan.steps = vec![
-                "vector_search",
-                "graph_expansion",
-                "context_pruning",
-                "global_fallback_snapshot_pinned",
-            ];
-            let mut state = self
-                .execute_with_plan(
-                    request,
-                    plan,
-                    embedding_model_id,
-                    snapshot_view,
-                    tenant_scope,
-                    session,
-                )
-                .await?;
-            state.exclusions.push(ExclusionReason {
-                node_id: None,
-                reason: "global_summary_disabled_by_snapshot_pin".to_string(),
-            });
-            return Ok((state, plan.clone(), None));
-        }
-
-        // Always build filtered evidence first so global synthesis can respect request filters.
-        let mut state = self
-            .execute_with_plan(
-                request,
-                plan,
-                embedding_model_id,
-                snapshot_view,
-                tenant_scope,
-                session,
-            )
-            .await?;
-
-        // MAP: Score community summaries.
-        let ranked = map_community_summaries(&request.query, &summary_candidates);
-        let relation_filter = collect_relation_filter(request);
-        let time_range = parse_time_range(request)?;
-        let retention_cutoff = retention_cutoff_unix(request);
-        let entity_filter: HashSet<&str> = request
-            .filters
-            .entity_type
-            .iter()
-            .map(|value| value.as_str())
-            .collect();
-
-        // Build lookup for top nodes referenced by summaries so we can enforce filters
-        // before producing global synthesis.
-        let all_top_node_ids: Vec<u64> = {
-            let mut seen = HashSet::new();
-            let mut out = Vec::new();
-            for (summary, _) in &ranked {
-                for node_id in &summary.top_nodes {
-                    if seen.insert(*node_id) {
-                        out.push(*node_id);
-                    }
-                }
-            }
-            out
-        };
-        let top_nodes = self
-            .get_nodes_by_ids_from_source(&all_top_node_ids, snapshot_view, None)
-            .await;
-        let top_node_lookup: HashMap<u64, Node> =
-            top_nodes.into_iter().map(|node| (node.id, node)).collect();
-
-        // Restrict summary candidates by relevance (>0) and request filters.
-        // Relation filters are edge/path-level constraints and cannot be validated
-        // purely from summary top nodes, so summary synthesis is disabled in that case.
-        let relevant_ranked: Vec<(&CommunitySummary, f32)> = if relation_filter.is_empty() {
-            ranked
-                .into_iter()
-                .filter(|(summary, score)| {
-                    *score > 0.0
-                        && summary.top_nodes.iter().any(|node_id| {
-                            top_node_lookup.get(node_id).is_some_and(|node| {
-                                node_passes_filters(
-                                    node,
-                                    &entity_filter,
-                                    time_range,
-                                    retention_cutoff,
-                                    tenant_scope,
-                                )
-                            })
-                        })
-                })
-                .collect()
-        } else {
-            Vec::new()
-        };
-
-        let max_communities = 5;
-        let global_answer = if relevant_ranked.is_empty() {
-            let reason = if relation_filter.is_empty() {
-                "global_no_relevant_community_summary"
-            } else {
-                "global_summary_disabled_by_relation_filter"
-            };
-            state.exclusions.push(ExclusionReason {
-                node_id: None,
-                reason: reason.to_string(),
-            });
-            None
-        } else {
-            Some(reduce_community_summaries(
-                &request.query,
-                &relevant_ranked,
-                max_communities,
-            ))
-        };
-
-        plan.steps = vec![
-            "vector_search",
-            "community_map_reduce",
-            "graph_expansion",
-            "context_pruning",
-        ];
-
-        Ok((state, plan.clone(), global_answer))
-    }
-
-    // -----------------------------------------------------------------------
-    // DRIFT Search: Dynamic Reasoning via Iterative Feedback and Traversals
-    // -----------------------------------------------------------------------
-    async fn execute_drift(
-        &self,
-        request: &QueryRequest,
-        plan: &mut QueryPlan,
-        embedding_model_id: &str,
-        snapshot_view: Option<&SnapshotView>,
-        tenant_scope: Option<&str>,
-        session: Option<&SessionGraph>,
-    ) -> Result<(ExecutionState, QueryPlan), QueryError> {
-        plan.effective_search_mode = SearchMode::Drift;
-
-        let mut best_state: Option<ExecutionState> = None;
-        let initial_depth = plan.expansion_depth;
-
-        for iteration in 0..DRIFT_MAX_ITERATIONS {
-            let mut iter_plan = plan.clone();
-            // Each DRIFT iteration expands the search depth progressively.
-            iter_plan.expansion_depth = (initial_depth + iteration as u8).min(8);
-            // Increase vector candidates each round.
-            iter_plan.vector_top_k = plan.vector_top_k.saturating_add(iteration * 2).min(50);
-
-            let state = self
-                .execute_with_plan(
-                    request,
-                    &iter_plan,
-                    embedding_model_id,
-                    snapshot_view,
-                    tenant_scope,
-                    session,
-                )
-                .await?;
-
-            let is_sufficient = state.nodes.len() >= DRIFT_EVIDENCE_THRESHOLD
-                || (iteration > 0
-                    && best_state
-                        .as_ref()
-                        .map(|prev| state.nodes.len() <= prev.nodes.len())
-                        .unwrap_or(false));
-
-            if state.nodes.len() > best_state.as_ref().map(|s| s.nodes.len()).unwrap_or(0) {
-                best_state = Some(state);
-            }
-
-            if is_sufficient {
-                break;
-            }
-        }
-
-        let mut state = best_state.unwrap_or(ExecutionState {
-            anchors: Vec::new(),
-            expansion_paths: Vec::new(),
-            exclusions: vec![ExclusionReason {
-                node_id: None,
-                reason: "drift_no_evidence_found".to_string(),
-            }],
-            nodes: Vec::new(),
-            edges: Vec::new(),
-        });
-
-        plan.steps = vec![
-            "vector_search",
-            "drift_iterative_expansion",
-            "graph_expansion",
-            "context_pruning",
-        ];
-
-        // If drift still yielded nothing, note it in exclusions.
-        if state.nodes.is_empty() {
-            state.exclusions.push(ExclusionReason {
-                node_id: None,
-                reason: "drift_exhausted_no_evidence".to_string(),
-            });
-        }
-
-        Ok((state, plan.clone()))
-    }
-
-    async fn execute_with_plan(
-        &self,
-        request: &QueryRequest,
-        plan: &QueryPlan,
-        embedding_model_id: &str,
-        snapshot_view: Option<&SnapshotView>,
-        tenant_scope: Option<&str>,
-        session: Option<&SessionGraph>,
-    ) -> Result<ExecutionState, QueryError> {
-        let mut vector_hits = self
-            .collect_vector_scores(
-                request,
-                plan,
-                embedding_model_id,
-                snapshot_view,
-                tenant_scope,
-                session,
-            )
-            .await;
-        if vector_hits.is_empty() {
-            if let Some(node_id) = self
-                .list_node_ids_from_source(snapshot_view, session)
-                .await
-                .into_iter()
-                .next()
-            {
-                vector_hits.push((node_id, 0.0));
-            }
-        }
-
-        if vector_hits.is_empty() {
-            return Ok(ExecutionState {
-                anchors: Vec::new(),
-                expansion_paths: Vec::new(),
-                exclusions: vec![ExclusionReason {
-                    node_id: None,
-                    reason: "no_nodes_available".to_string(),
-                }],
-                nodes: Vec::new(),
-                edges: Vec::new(),
-            });
-        }
-
-        let anchor_limit = plan.vector_top_k.min(vector_hits.len()).max(1);
-        let mut anchors: Vec<Anchor> = vector_hits
-            .iter()
-            .take(anchor_limit)
-            .map(|(node_id, score)| Anchor {
-                node_id: *node_id,
-                score: *score,
-            })
-            .collect();
-
-        let relation_filter = collect_relation_filter(request);
-        let mut candidate_hops: HashMap<u64, u8> = HashMap::new();
-        let mut expansion_paths = Vec::new();
-        let mut exclusions = Vec::new();
-        let mut traversed_edges = Vec::new();
-
-        if let Some(view) = snapshot_view {
-            for anchor in &anchors {
-                candidate_hops.entry(anchor.node_id).or_insert(0);
-
-                let mut queue = VecDeque::new();
-                let mut visited: HashMap<u64, u8> = HashMap::new();
-                let mut parents: HashMap<u64, u64> = HashMap::new();
-
-                queue.push_back(anchor.node_id);
-                visited.insert(anchor.node_id, 0);
-
-                while let Some(current_id) = queue.pop_front() {
-                    let current_hop = *visited.get(&current_id).unwrap_or(&0);
-                    if current_hop >= plan.expansion_depth {
-                        continue;
-                    }
-
-                    for (target, relation, weight) in
-                        view.neighbors_with_session(current_id, session)
-                    {
-                        if !relation_is_allowed(relation.as_str(), &relation_filter) {
-                            exclusions.push(ExclusionReason {
-                                node_id: Some(target),
-                                reason: format!("relation_filtered:{}", relation),
-                            });
-                            continue;
-                        }
-
-                        traversed_edges.push(InternalEdge {
-                            source: current_id,
-                            target,
-                            relation: relation.clone(),
-                            weight,
-                            provenance: Provenance::default(),
-                            confidence: weight,
-                        });
-
-                        let next_hop = current_hop + 1;
-                        let should_visit = visited
-                            .get(&target)
-                            .map(|prev_hop| next_hop < *prev_hop)
-                            .unwrap_or(true);
-
-                        if should_visit {
-                            visited.insert(target, next_hop);
-                            parents.insert(target, current_id);
-                            queue.push_back(target);
-                            candidate_hops
-                                .entry(target)
-                                .and_modify(|hop| *hop = (*hop).min(next_hop))
-                                .or_insert(next_hop);
-
-                            if let Some(path) = reconstruct_path(anchor.node_id, target, &parents) {
-                                expansion_paths.push(ExpansionPath {
-                                    anchor_id: anchor.node_id,
-                                    target_id: target,
-                                    path,
-                                });
-                            }
-                        }
-                    }
-                }
-            }
-        } else {
-            for anchor in &anchors {
-                candidate_hops.entry(anchor.node_id).or_insert(0);
-
-                let mut queue = VecDeque::new();
-                let mut visited: HashMap<u64, u8> = HashMap::new();
-                let mut parents: HashMap<u64, u64> = HashMap::new();
-
-                queue.push_back(anchor.node_id);
-                visited.insert(anchor.node_id, 0);
-
-                while let Some(current_id) = queue.pop_front() {
-                    let current_hop = *visited.get(&current_id).unwrap_or(&0);
-                    if current_hop >= plan.expansion_depth {
-                        continue;
-                    }
-
-                    for (target, relation, weight) in self
-                        .repo
-                        .neighbors_with_session_graph(current_id, session)
-                        .await
-                    {
-                        if !relation_is_allowed(relation.as_str(), &relation_filter) {
-                            exclusions.push(ExclusionReason {
-                                node_id: Some(target),
-                                reason: format!("relation_filtered:{}", relation),
-                            });
-                            continue;
-                        }
-
-                        traversed_edges.push(InternalEdge {
-                            source: current_id,
-                            target,
-                            relation: relation.clone(),
-                            weight,
-                            provenance: Provenance::default(),
-                            confidence: weight,
-                        });
-
-                        let next_hop = current_hop + 1;
-                        let should_visit = visited
-                            .get(&target)
-                            .map(|prev_hop| next_hop < *prev_hop)
-                            .unwrap_or(true);
-
-                        if should_visit {
-                            visited.insert(target, next_hop);
-                            parents.insert(target, current_id);
-                            queue.push_back(target);
-                            candidate_hops
-                                .entry(target)
-                                .and_modify(|hop| *hop = (*hop).min(next_hop))
-                                .or_insert(next_hop);
-
-                            if let Some(path) = reconstruct_path(anchor.node_id, target, &parents) {
-                                expansion_paths.push(ExpansionPath {
-                                    anchor_id: anchor.node_id,
-                                    target_id: target,
-                                    path,
-                                });
-                            }
-                        }
-                    }
-                }
-            }
-        }
-
-        let candidate_ids: Vec<u64> = candidate_hops.keys().copied().collect();
-        let fetched_nodes = self
-            .get_nodes_by_ids_from_source(&candidate_ids, snapshot_view, session)
-            .await;
-        let node_lookup: HashMap<u64, Node> = fetched_nodes
-            .into_iter()
-            .map(|node| (node.id, node))
-            .collect();
-
-        let anchor_scores: HashMap<u64, f32> = anchors
-            .iter()
-            .map(|anchor| (anchor.node_id, anchor.score))
-            .collect();
-        let query_tokens = tokenize(&request.query);
-        let time_range = parse_time_range(request)?;
-        let retention_cutoff = retention_cutoff_unix(request);
-        let entity_filter: HashSet<&str> = request
-            .filters
-            .entity_type
-            .iter()
-            .map(|value| value.as_str())
-            .collect();
-
-        let mut ranked_nodes = Vec::new();
-        for (node_id, hop) in candidate_hops {
-            let Some(node) = node_lookup.get(&node_id) else {
-                exclusions.push(ExclusionReason {
-                    node_id: Some(node_id),
-                    reason: "missing_node".to_string(),
-                });
-                continue;
-            };
-
-            if let Some(reason) = node_filter_exclusion_reason(
-                node,
-                &entity_filter,
-                time_range,
-                retention_cutoff,
-                tenant_scope,
-            ) {
-                exclusions.push(ExclusionReason {
-                    node_id: Some(node_id),
-                    reason,
-                });
-                continue;
-            }
-
-            let lexical_score =
-                lexical_similarity(&query_tokens, &tokenize(&node_lexical_text(node)));
-            let anchor_score = anchor_scores.get(&node_id).copied().unwrap_or(0.0);
-            let base_score = ((anchor_score * 0.8) + (lexical_score * 0.2))
-                .max(lexical_score)
-                .max(0.01);
-            let score = base_score / (hop as f32 + 1.0);
-
-            // Extract confidence: prefer explicit metadata, fall back to score
-            let confidence = node
-                .metadata
-                .get("confidence")
-                .and_then(|v| v.parse::<f32>().ok())
-                .unwrap_or(score);
-
-            ranked_nodes.push(RankedNode {
-                id: node_id,
-                data: node.data.clone(),
-                score,
-                hop,
-                source: node.metadata.get("source").cloned(),
-                extraction_model_id: node.metadata.get("extraction_model_id").cloned(),
-                node_snapshot_id: node.metadata.get("snapshot_id").cloned(),
-                ingested_at: node.metadata.get("ingested_at").cloned(),
-                confidence,
-            });
-        }
-
-        ranked_nodes.sort_by(|a, b| {
-            b.score
-                .partial_cmp(&a.score)
-                .unwrap_or(Ordering::Equal)
-                .then(a.id.cmp(&b.id))
-        });
-
-        if ranked_nodes.len() > request.top_k {
-            let pruned = ranked_nodes.split_off(request.top_k);
-            for node in pruned {
-                exclusions.push(ExclusionReason {
-                    node_id: Some(node.id),
-                    reason: "pruned_by_top_k".to_string(),
-                });
-            }
-        }
-
-        let selected_ids: HashSet<u64> = ranked_nodes.iter().map(|node| node.id).collect();
-
-        let mut edges: Vec<InternalEdge> = traversed_edges
-            .into_iter()
-            .filter(|edge| {
-                selected_ids.contains(&edge.source) && selected_ids.contains(&edge.target)
-            })
-            .filter(|edge| relation_is_allowed(edge.relation.as_str(), &relation_filter))
-            .collect();
-
-        // Enrich edges with provenance — single lock acquisition via bulk API
-        {
-            let edge_keys: Vec<(u64, u64, String)> = edges
-                .iter()
-                .map(|e| (e.source, e.target, e.relation.clone()))
-                .collect();
-            let all_meta = self
-                .get_edge_metadata_bulk_from_source(&edge_keys, snapshot_view)
-                .await;
-            for edge in &mut edges {
-                let key = (edge.source, edge.target, edge.relation.clone());
-                if let Some(meta) = all_meta.get(&key) {
-                    edge.provenance = Provenance {
-                        source: meta.get("source").cloned(),
-                        extraction_model_id: meta.get("extraction_model_id").cloned(),
-                        snapshot_id: meta.get("snapshot_id").cloned(),
-                        ingested_at: meta.get("ingested_at").cloned(),
-                    };
-                }
-            }
-        }
-
-        edges = dedup_edges(edges);
-        expansion_paths = dedup_paths(expansion_paths);
-        exclusions = dedup_exclusions(exclusions);
-        anchors.sort_by(|a, b| {
-            b.score
-                .partial_cmp(&a.score)
-                .unwrap_or(Ordering::Equal)
-                .then(a.node_id.cmp(&b.node_id))
-        });
-
-        Ok(ExecutionState {
-            anchors,
-            expansion_paths,
-            exclusions,
-            nodes: ranked_nodes,
-            edges,
-        })
-    }
-
-    async fn collect_vector_scores(
-        &self,
-        request: &QueryRequest,
-        plan: &QueryPlan,
-        embedding_model_id: &str,
-        snapshot_view: Option<&SnapshotView>,
-        tenant_scope: Option<&str>,
-        session: Option<&SessionGraph>,
-    ) -> Vec<(u64, f32)> {
-        let embedding_dim = match snapshot_view {
-            Some(view) => view.embedding_dimension(),
-            None => self.repo.embedding_dimension().await,
-        }
-        .or_else(|| session.and_then(SessionGraph::embedding_dimension));
-        let Some(embedding_dim) = embedding_dim else {
-            return Vec::new();
-        };
-
-        let query_embedding =
-            deterministic_embedding(&request.query, embedding_model_id, embedding_dim);
-        let vector_limit = match plan.effective_search_mode {
-            SearchMode::Global => plan.vector_top_k.saturating_mul(2),
-            _ => plan.vector_top_k,
-        }
-        .max(1);
-
-        let raw_hits = match snapshot_view {
-            Some(view) => view.search_vector_with_session(&query_embedding, vector_limit, session),
-            None => {
-                self.repo
-                    .search_vector_with_session_graph(&query_embedding, vector_limit, session)
-                    .await
-            }
-        };
-
-        let Some(tenant) = tenant_scope else {
-            return raw_hits;
-        };
-
-        let candidate_ids: Vec<u64> = raw_hits.iter().map(|(node_id, _)| *node_id).collect();
-        let allowed_ids: HashSet<u64> = self
-            .get_nodes_by_ids_from_source(&candidate_ids, snapshot_view, session)
-            .await
-            .into_iter()
-            .filter(|node| node_belongs_to_tenant(node, tenant))
-            .map(|node| node.id)
-            .collect();
-
-        raw_hits
-            .into_iter()
-            .filter(|(node_id, _)| allowed_ids.contains(node_id))
-            .collect()
-    }
-}
-
-fn node_belongs_to_tenant(node: &Node, tenant_scope: &str) -> bool {
-    node.metadata
-        .get("tenant")
-        .map(|tenant| tenant == tenant_scope)
-        .unwrap_or(false)
-}
-
-fn node_lexical_text(node: &Node) -> String {
-    format!(
-        "{} {}",
-        node.data,
-        node.metadata
-            .values()
-            .cloned()
-            .collect::<Vec<_>>()
-            .join(" ")
-    )
-}
-
-fn tokenize(text: &str) -> HashSet<String> {
-    let mut out = HashSet::new();
-    let mut buffer = String::new();
-
-    for ch in text.chars().flat_map(|ch| ch.to_lowercase()) {
-        if ch.is_alphanumeric() || ch == '_' {
-            buffer.push(ch);
-        } else if !buffer.is_empty() {
-            out.insert(buffer.clone());
-            buffer.clear();
-        }
-    }
-
-    if !buffer.is_empty() {
-        out.insert(buffer);
-    }
-
-    let unicode_tokens: Vec<String> = out
-        .iter()
-        .filter(|token| !token.is_ascii())
-        .cloned()
-        .collect();
-    for token in unicode_tokens {
-        for ngram in char_ngrams(&token, UNICODE_NGRAM_SIZE) {
-            out.insert(ngram);
-        }
-    }
-
-    out
-}
-
-fn char_ngrams(token: &str, n: usize) -> Vec<String> {
-    let chars: Vec<char> = token.chars().collect();
-    if chars.is_empty() || n == 0 {
-        return Vec::new();
-    }
-    if chars.len() <= n {
-        return vec![token.to_string()];
-    }
-
-    chars
-        .windows(n)
-        .map(|window| window.iter().collect::<String>())
-        .collect()
-}
-
-fn lexical_similarity(a: &HashSet<String>, b: &HashSet<String>) -> f32 {
-    if a.is_empty() || b.is_empty() {
-        return 0.0;
-    }
-
-    let intersection = a.intersection(b).count() as f32;
-    let denominator = a.len().max(b.len()) as f32;
-    intersection / denominator
-}
-
-fn collect_relation_filter(request: &QueryRequest) -> HashSet<&str> {
-    request
-        .filters
-        .relation_type
-        .iter()
-        .map(|value| value.as_str())
-        .chain(
-            request
-                .traversal
-                .relation_types
-                .iter()
-                .map(|value| value.as_str()),
-        )
-        .collect()
-}
-
-fn relation_is_allowed(relation: &str, relation_filter: &HashSet<&str>) -> bool {
-    relation_filter.is_empty() || relation_filter.contains(relation)
-}
-
-fn reconstruct_path(
-    anchor_id: u64,
-    target_id: u64,
-    parents: &HashMap<u64, u64>,
-) -> Option<Vec<u64>> {
-    let mut path = vec![target_id];
-    let mut current = target_id;
-
-    while current != anchor_id {
-        let parent = parents.get(&current)?;
-        current = *parent;
-        path.push(current);
-    }
-
-    path.reverse();
-    Some(path)
-}
-
-fn parse_time_range(request: &QueryRequest) -> Result<Option<(NaiveDate, NaiveDate)>, QueryError> {
-    let Some(range) = &request.filters.time_range else {
-        return Ok(None);
-    };
-
-    let from = NaiveDate::parse_from_str(&range.from, "%Y-%m-%d").map_err(|_| {
-        QueryError::InvalidQuery("filters.time_range.from/to must be YYYY-MM-DD".to_string())
-    })?;
-    let to = NaiveDate::parse_from_str(&range.to, "%Y-%m-%d").map_err(|_| {
-        QueryError::InvalidQuery("filters.time_range.from/to must be YYYY-MM-DD".to_string())
-    })?;
-
-    Ok(Some((from, to)))
-}
-
-fn parse_time_travel_as_of_unix_ms(input: &str) -> Result<i64, QueryError> {
-    if let Ok(date) = NaiveDate::parse_from_str(input, "%Y-%m-%d") {
-        return date
-            .and_hms_milli_opt(23, 59, 59, 999)
-            .map(|datetime| datetime.and_utc().timestamp_millis())
-            .ok_or_else(|| {
-                QueryError::InvalidQuery(
-                    "time_travel must be YYYY-MM-DD or RFC3339 format".to_string(),
-                )
-            });
-    }
-
-    DateTime::parse_from_rfc3339(input)
-        .map(|datetime| datetime.with_timezone(&Utc).timestamp_millis())
-        .map_err(|_| {
-            QueryError::InvalidQuery("time_travel must be YYYY-MM-DD or RFC3339 format".to_string())
-        })
-}
-
-fn node_filter_exclusion_reason(
-    node: &Node,
-    entity_filter: &HashSet<&str>,
-    time_range: Option<(NaiveDate, NaiveDate)>,
-    retention_cutoff_unix: Option<u64>,
-    tenant_scope: Option<&str>,
-) -> Option<String> {
-    if let Some(tenant_scope) = tenant_scope {
-        if !node_belongs_to_tenant(node, tenant_scope) {
-            return Some("tenant_filtered".to_string());
-        }
-    }
-
-    if let Some(now_unix) = retention_cutoff_unix {
-        if node_is_retention_expired(node, now_unix) {
-            return Some("retention_expired".to_string());
-        }
-    }
-
-    if !entity_filter.is_empty() {
-        let entity_type = node.metadata.get("entity_type").map(|value| value.as_str());
-        if entity_type
-            .map(|value| !entity_filter.contains(value))
-            .unwrap_or(true)
-        {
-            return Some("entity_type_filtered".to_string());
-        }
-    }
-
-    if let Some((from, to)) = time_range {
-        let timestamp = node
-            .metadata
-            .get("timestamp")
-            .and_then(|value| NaiveDate::parse_from_str(value, "%Y-%m-%d").ok());
-        match timestamp {
-            Some(value) if value >= from && value <= to => {}
-            _ => return Some("time_range_filtered".to_string()),
-        }
-    }
-
-    None
-}
-
-fn node_passes_filters(
-    node: &Node,
-    entity_filter: &HashSet<&str>,
-    time_range: Option<(NaiveDate, NaiveDate)>,
-    retention_cutoff_unix: Option<u64>,
-    tenant_scope: Option<&str>,
-) -> bool {
-    node_filter_exclusion_reason(
-        node,
-        entity_filter,
-        time_range,
-        retention_cutoff_unix,
-        tenant_scope,
-    )
-    .is_none()
-}
-
-fn node_is_retention_expired(node: &Node, now_unix: u64) -> bool {
-    node.metadata
-        .get("retention_until_unix")
-        .and_then(|raw| raw.parse::<u64>().ok())
-        .is_some_and(|deadline| now_unix >= deadline)
-}
-
-fn current_unix_timestamp() -> u64 {
-    std::time::SystemTime::now()
-        .duration_since(std::time::UNIX_EPOCH)
-        .unwrap_or_default()
-        .as_secs()
-}
-
-fn retention_cutoff_unix(request: &QueryRequest) -> Option<u64> {
-    if request.snapshot_id.is_some() || request.time_travel.is_some() {
-        None
-    } else {
-        Some(current_unix_timestamp())
-    }
-}
-
-fn effective_query_model_id(request: &QueryRequest) -> String {
-    request
-        .model_id
-        .clone()
-        .unwrap_or_else(|| DEFAULT_EMBEDDING_MODEL_ID.to_string())
-}
-
-fn build_query_audit_event(
-    outcome: AuditOutcome,
-    model_id: &str,
-    actor: Option<String>,
-    tenant: Option<String>,
-    snapshot_id: Option<String>,
-    error: Option<String>,
-) -> AuditEvent {
-    let mut event = AuditEvent::new(AuditOperation::Query, outcome);
-    event.model_id = Some(model_id.to_string());
-    event.actor = actor;
-    event.tenant = tenant;
-    event.snapshot_id = snapshot_id;
-    if let Some(error) = error {
-        event.metadata.insert("error".to_string(), error);
-    }
-    event
-}
-
-fn generate_answer(query: &str, nodes: &[EvidenceNode]) -> String {
-    if nodes.is_empty() {
-        return format!("No evidence found for query: {query}");
-    }
-
-    let snippets = nodes
-        .iter()
-        .take(3)
-        .map(|node| node.data.as_str())
-        .collect::<Vec<_>>()
-        .join(" | ");
-
-    format!(
-        "Answer synthesized from {} evidence nodes: {}",
-        nodes.len(),
-        snippets
-    )
-}
-
-fn build_citations(nodes: &[RankedNode]) -> Vec<Citation> {
-    let mut unique_sources = BTreeSet::new();
-    let mut out = Vec::new();
-
-    for node in nodes {
-        let Some(source) = &node.source else {
-            continue;
-        };
-        if !unique_sources.insert(source.clone()) {
-            continue;
-        }
-
-        let end = node.data.len();
-        out.push(Citation {
-            source: source.clone(),
-            span: [0, end],
-            node_id: node.id,
-            confidence: node.confidence,
-        });
-    }
-
-    out
-}
-
-fn dedup_edges(edges: Vec<InternalEdge>) -> Vec<InternalEdge> {
-    // Last-wins dedup: when the same (source, target, relation) appears multiple times,
-    // keep the last occurrence which has the most up-to-date weight/provenance.
-    let mut map: HashMap<(u64, u64, String), InternalEdge> = HashMap::new();
-    for edge in edges {
-        let key = (edge.source, edge.target, edge.relation.clone());
-        map.insert(key, edge);
-    }
-    let mut out: Vec<InternalEdge> = map.into_values().collect();
-    out.sort_by(|a, b| {
-        a.source
-            .cmp(&b.source)
-            .then(a.target.cmp(&b.target))
-            .then(a.relation.cmp(&b.relation))
-    });
-    out
-}
-
-fn dedup_paths(mut paths: Vec<ExpansionPath>) -> Vec<ExpansionPath> {
-    let mut seen = BTreeSet::new();
-    paths.retain(|path| seen.insert((path.anchor_id, path.target_id, path.path.clone())));
-    paths.sort_by(|a, b| {
-        a.anchor_id
-            .cmp(&b.anchor_id)
-            .then(a.target_id.cmp(&b.target_id))
-    });
-    paths
-}
-
-fn dedup_exclusions(mut exclusions: Vec<ExclusionReason>) -> Vec<ExclusionReason> {
-    let mut seen = BTreeSet::new();
-    exclusions.retain(|exclusion| seen.insert((exclusion.node_id, exclusion.reason.clone())));
-    exclusions.sort_by(|a, b| a.node_id.cmp(&b.node_id).then(a.reason.cmp(&b.reason)));
-    exclusions
-}
