--- storage/src/repo.rs
@@ -1,2121 +1,0 @@
-use crate::crypto::{AtRestCipher, NoOpCipher};
-use crate::hyper_index::HyperIndex;
-use crate::index::AdjacencyGraph;
-use crate::session::{SessionGraph, SessionManager, SessionOwner};
-use crate::snapshot::{SnapshotCatalog, SnapshotCatalogEntry, SnapshotError, SnapshotManager};
-use crate::tiering::{StorageCapabilities, StorageProfile};
-use crate::wal::{Wal, WalError, WalOptions};
-use alayasiki_core::error::{AlayasikiError, ErrorCode};
-use alayasiki_core::model::{Edge, Node};
-use rkyv::ser::{serializers::AllocSerializer, Serializer};
-use rkyv::{Archive, Deserialize, Serialize};
-use std::collections::{HashMap, HashSet};
-use std::path::{Path, PathBuf};
-use std::sync::Arc;
-use std::time::Duration;
-use thiserror::Error;
-use tokio::fs;
-use tokio::sync::{Mutex, RwLock};
-
-#[derive(Error, Debug)]
-pub enum RepoError {
-    #[error("WAL error: {0}")]
-    Wal(#[from] WalError),
-    #[error("Serialization error")]
-    Serialization,
-    #[error("Deserialization error")]
-    Deserialization,
-    #[error("Not found")]
-    NotFound,
-    #[error("Invalid transaction: {0}")]
-    InvalidTransaction(String),
-    #[error("Invalid snapshot id: {0}")]
-    InvalidSnapshotId(String),
-    #[error("Snapshot not found: {0}")]
-    SnapshotNotFound(String),
-    #[error("Snapshot manager is not configured")]
-    SnapshotNotConfigured,
-    #[error("Snapshot error: {0}")]
-    Snapshot(#[from] SnapshotError),
-    #[error("Session access denied: {0}")]
-    SessionAccessDenied(String),
-}
-
-impl AlayasikiError for RepoError {
-    fn error_code(&self) -> ErrorCode {
-        match self {
-            RepoError::Wal(err) => err.error_code(),
-            RepoError::Serialization => ErrorCode::Internal,
-            RepoError::Deserialization => ErrorCode::Internal,
-            RepoError::NotFound => ErrorCode::NotFound,
-            RepoError::InvalidTransaction(_) => ErrorCode::InvalidArgument,
-            RepoError::InvalidSnapshotId(_) => ErrorCode::InvalidArgument,
-            RepoError::SnapshotNotFound(_) => ErrorCode::NotFound,
-            RepoError::SnapshotNotConfigured => ErrorCode::Internal,
-            RepoError::Snapshot(err) => err.error_code(),
-            RepoError::SessionAccessDenied(_) => ErrorCode::PermissionDenied,
-        }
-    }
-}
-
-/// WAL Entry types for durability
-#[derive(Archive, Deserialize, Serialize, Debug, Clone)]
-#[archive(check_bytes)]
-pub enum WalEntry {
-    Put(Node),
-    PutEdge(Edge),
-    Delete(u64),
-    IdempotencyKey { key: String, node_ids: Vec<u64> },
-    Transaction(Vec<TxOperation>),
-}
-
-#[derive(Archive, Deserialize, Serialize, Debug, Clone)]
-#[archive(check_bytes)]
-pub enum TxOperation {
-    Put(Node),
-    PutEdge(Edge),
-    Delete(u64),
-    RecordIdempotency { key: String, node_ids: Vec<u64> },
-}
-
-#[derive(Debug, Clone)]
-pub enum IndexMutation {
-    PutNode(Node),
-    PutEdge(Edge),
-    DeleteNode(u64),
-}
-
-/// Key for edge metadata lookup: (source, target, relation)
-pub type EdgeMetaKey = (u64, u64, String);
-
-#[derive(Archive, Deserialize, Serialize, Debug, Clone)]
-#[archive(check_bytes)]
-struct BackupEdgeRecord {
-    source: u64,
-    target: u64,
-    relation: String,
-    weight: f32,
-}
-
-#[derive(Archive, Deserialize, Serialize, Debug, Clone)]
-#[archive(check_bytes)]
-struct BackupIdempotencyRecord {
-    key: String,
-    node_ids: Vec<u64>,
-}
-
-#[derive(Archive, Deserialize, Serialize, Debug, Clone)]
-#[archive(check_bytes)]
-struct BackupEdgeMetadataRecord {
-    source: u64,
-    target: u64,
-    relation: String,
-    metadata: HashMap<String, String>,
-}
-
-#[derive(Archive, Deserialize, Serialize, Debug, Clone)]
-#[archive(check_bytes)]
-struct RepositoryBackupSnapshot {
-    lsn: u64,
-    nodes: Vec<Node>,
-    edges: Vec<BackupEdgeRecord>,
-    idempotency: Vec<BackupIdempotencyRecord>,
-    edge_metadata: Vec<BackupEdgeMetadataRecord>,
-}
-
-struct MaterializedState {
-    nodes: HashMap<u64, Node>,
-    hyper_index: HyperIndex,
-    idempotency_index: HashMap<String, Vec<u64>>,
-    edge_metadata: HashMap<EdgeMetaKey, HashMap<String, String>>,
-}
-
-pub struct SnapshotView {
-    snapshot_id: String,
-    nodes: HashMap<u64, Node>,
-    hyper_index: HyperIndex,
-    edge_metadata: HashMap<EdgeMetaKey, HashMap<String, String>>,
-}
-
-impl SnapshotView {
-    pub fn snapshot_id(&self) -> &str {
-        &self.snapshot_id
-    }
-
-    pub fn storage_capabilities(&self) -> &StorageCapabilities {
-        self.hyper_index.storage_capabilities()
-    }
-
-    pub fn list_node_ids(&self) -> Vec<u64> {
-        let mut out: Vec<u64> = self.nodes.keys().copied().collect();
-        out.sort_unstable();
-        out
-    }
-
-    pub fn get_nodes_by_ids(&self, ids: &[u64]) -> Vec<Node> {
-        let mut out: Vec<Node> = ids
-            .iter()
-            .filter_map(|id| self.nodes.get(id).cloned())
-            .collect();
-        out.sort_by_key(|node| node.id);
-        out
-    }
-
-    pub fn embedding_dimension(&self) -> Option<usize> {
-        self.nodes
-            .values()
-            .find_map(|node| (!node.embedding.is_empty()).then_some(node.embedding.len()))
-    }
-
-    pub fn search_vector(&self, query: &[f32], k: usize) -> Vec<(u64, f32)> {
-        self.hyper_index.search_vector(query, k)
-    }
-
-    pub fn search_vector_with_session(
-        &self,
-        query: &[f32],
-        k: usize,
-        session: Option<&crate::session::SessionGraph>,
-    ) -> Vec<(u64, f32)> {
-        let mut results = self.search_vector(query, k);
-        if let Some(session) = session {
-            use alayasiki_core::embedding::cosine_similarity;
-            let mut session_results: Vec<(u64, f32)> = session
-                .nodes
-                .values()
-                .filter_map(|node| {
-                    cosine_similarity(query, &node.embedding).map(|sim| (node.id, sim))
-                })
-                .collect();
-
-            results.append(&mut session_results);
-            results.sort_by(|a, b| {
-                a.0.cmp(&b.0)
-                    .then_with(|| b.1.partial_cmp(&a.1).unwrap_or(std::cmp::Ordering::Equal))
-            });
-            results.dedup_by_key(|(id, _)| *id);
-            results.sort_by(|a, b| b.1.partial_cmp(&a.1).unwrap_or(std::cmp::Ordering::Equal));
-            results.truncate(k);
-        }
-        results
-    }
-
-    pub fn neighbors(&self, node_id: u64) -> Vec<(u64, String, f32)> {
-        self.hyper_index
-            .graph_index
-            .neighbors(node_id)
-            .into_iter()
-            .cloned()
-            .collect()
-    }
-
-    pub fn neighbors_with_session(
-        &self,
-        node_id: u64,
-        session: Option<&crate::session::SessionGraph>,
-    ) -> Vec<(u64, String, f32)> {
-        let mut results = self.neighbors(node_id);
-        if let Some(session) = session {
-            for edge in &session.edges {
-                if edge.source == node_id {
-                    results.push((edge.target, edge.relation.clone(), edge.weight));
-                }
-            }
-        }
-        results
-    }
-
-    pub fn get_edge_metadata_bulk(
-        &self,
-        keys: &[(u64, u64, String)],
-    ) -> HashMap<EdgeMetaKey, HashMap<String, String>> {
-        keys.iter()
-            .filter_map(|key| {
-                self.edge_metadata
-                    .get(key)
-                    .map(|meta| (key.clone(), meta.clone()))
-            })
-            .collect()
-    }
-}
-
-pub struct Repository {
-    wal: Arc<Mutex<Wal>>,
-    tx_lock: Arc<Mutex<()>>,
-    nodes: Arc<RwLock<HashMap<u64, Node>>>,
-    pub hyper_index: Arc<RwLock<HyperIndex>>,
-    idempotency_index: Arc<RwLock<HashMap<String, Vec<u64>>>>,
-    edge_metadata: Arc<RwLock<HashMap<EdgeMetaKey, HashMap<String, String>>>>,
-    snapshot_manager: Option<SnapshotManager>,
-    snapshot_catalog: Arc<Mutex<SnapshotCatalog>>,
-    pub session_manager: Arc<SessionManager>,
-    storage_profile: StorageProfile,
-    storage_capabilities: StorageCapabilities,
-}
-
-const DEFAULT_SESSION_TTL: Duration = Duration::from_secs(30 * 60);
-
-impl Repository {
-    /// Create a new empty Repository (no replay)
-    pub fn new(wal: Arc<Mutex<Wal>>) -> Self {
-        Self::new_with_profile(wal, StorageProfile::default())
-    }
-
-    pub fn new_with_profile(wal: Arc<Mutex<Wal>>, storage_profile: StorageProfile) -> Self {
-        let storage_capabilities = storage_profile.resolve_capabilities();
-
-        Self {
-            wal,
-            tx_lock: Arc::new(Mutex::new(())),
-            nodes: Arc::new(RwLock::new(HashMap::new())),
-            hyper_index: Arc::new(RwLock::new(HyperIndex::with_storage_profile(
-                storage_profile.clone(),
-            ))),
-            idempotency_index: Arc::new(RwLock::new(HashMap::new())),
-            edge_metadata: Arc::new(RwLock::new(HashMap::new())),
-            snapshot_manager: None,
-            snapshot_catalog: Arc::new(Mutex::new(SnapshotCatalog::new_in_memory())),
-            session_manager: Arc::new(SessionManager::new(DEFAULT_SESSION_TTL)),
-            storage_profile,
-            storage_capabilities,
-        }
-    }
-
-    /// Open a Repository with WAL replay to restore previous state
-    pub async fn open(wal_path: impl AsRef<Path>) -> Result<Self, RepoError> {
-        Self::open_with_profile_and_options(
-            wal_path,
-            StorageProfile::default(),
-            WalOptions::default(),
-        )
-        .await
-    }
-
-    /// Open a repository with custom WAL recovery and flush options.
-    pub async fn open_with_options(
-        wal_path: impl AsRef<Path>,
-        wal_options: WalOptions,
-    ) -> Result<Self, RepoError> {
-        Self::open_with_profile_and_options(wal_path, StorageProfile::default(), wal_options).await
-    }
-
-    pub async fn open_with_profile(
-        wal_path: impl AsRef<Path>,
-        storage_profile: StorageProfile,
-    ) -> Result<Self, RepoError> {
-        Self::open_with_profile_and_options(wal_path, storage_profile, WalOptions::default()).await
-    }
-
-    pub async fn open_with_profile_and_options(
-        wal_path: impl AsRef<Path>,
-        storage_profile: StorageProfile,
-        wal_options: WalOptions,
-    ) -> Result<Self, RepoError> {
-        Self::open_internal(
-            wal_path.as_ref().to_path_buf(),
-            Arc::new(NoOpCipher),
-            None,
-            wal_options,
-            storage_profile,
-        )
-        .await
-    }
-
-    /// Open a repository with a custom at-rest cipher for WAL replay and writes.
-    pub async fn open_with_cipher(
-        wal_path: impl AsRef<Path>,
-        cipher: Arc<dyn AtRestCipher>,
-    ) -> Result<Self, RepoError> {
-        Self::open_with_cipher_and_options(wal_path, cipher, WalOptions::default()).await
-    }
-
-    /// Open a repository with custom at-rest cipher and WAL options.
-    pub async fn open_with_cipher_and_options(
-        wal_path: impl AsRef<Path>,
-        cipher: Arc<dyn AtRestCipher>,
-        wal_options: WalOptions,
-    ) -> Result<Self, RepoError> {
-        Self::open_internal(
-            wal_path.as_ref().to_path_buf(),
-            cipher,
-            None,
-            wal_options,
-            StorageProfile::default(),
-        )
-        .await
-    }
-
-    /// Open a repository and restore state from backup snapshots first, then WAL deltas.
-    pub async fn open_with_snapshots(
-        wal_path: impl AsRef<Path>,
-        snapshot_dir: impl AsRef<Path>,
-    ) -> Result<Self, RepoError> {
-        Self::open_with_cipher_and_snapshots_and_options(
-            wal_path,
-            Arc::new(NoOpCipher),
-            snapshot_dir,
-            WalOptions::default(),
-        )
-        .await
-    }
-
-    /// Open a repository with custom cipher and snapshot-backed recovery.
-    pub async fn open_with_cipher_and_snapshots(
-        wal_path: impl AsRef<Path>,
-        cipher: Arc<dyn AtRestCipher>,
-        snapshot_dir: impl AsRef<Path>,
-    ) -> Result<Self, RepoError> {
-        Self::open_with_cipher_and_snapshots_and_options(
-            wal_path,
-            cipher,
-            snapshot_dir,
-            WalOptions::default(),
-        )
-        .await
-    }
-
-    /// Open a repository with custom cipher, snapshot-backed recovery, and WAL options.
-    pub async fn open_with_cipher_and_snapshots_and_options(
-        wal_path: impl AsRef<Path>,
-        cipher: Arc<dyn AtRestCipher>,
-        snapshot_dir: impl AsRef<Path>,
-        wal_options: WalOptions,
-    ) -> Result<Self, RepoError> {
-        let snapshot_manager = SnapshotManager::new(snapshot_dir.as_ref());
-        Self::open_internal(
-            wal_path.as_ref().to_path_buf(),
-            cipher,
-            Some(snapshot_manager),
-            wal_options,
-            StorageProfile::default(),
-        )
-        .await
-    }
-
-    async fn open_internal(
-        wal_path: PathBuf,
-        cipher: Arc<dyn AtRestCipher>,
-        snapshot_manager: Option<SnapshotManager>,
-        wal_options: WalOptions,
-        storage_profile: StorageProfile,
-    ) -> Result<Self, RepoError> {
-        let wal_instance =
-            Wal::open_with_cipher_and_options(&wal_path, cipher, wal_options).await?;
-        let wal = Arc::new(Mutex::new(wal_instance));
-        let tx_lock = Arc::new(Mutex::new(()));
-        let (mut materialized, base_lsn) = load_materialized_state_from_backup(
-            snapshot_manager.as_ref(),
-            None,
-            storage_profile.clone(),
-        )
-        .await?;
-
-        // Replay WAL entries newer than the snapshot baseline.
-        {
-            let mut wal_lock = wal.lock().await;
-            let last_replayed_lsn = wal_lock
-                .replay(|lsn, data| {
-                    if lsn <= base_lsn {
-                        return Ok(());
-                    }
-                    // Deserialize (zero-copy check)
-                    let archived = rkyv::check_archived_root::<WalEntry>(&data[..])
-                        .map_err(|_| WalError::CorruptEntry)?;
-                    let entry: WalEntry = archived.deserialize(&mut rkyv::Infallible).unwrap();
-                    apply_replayed_entry(
-                        &entry,
-                        &mut materialized.nodes,
-                        &mut materialized.hyper_index,
-                        &mut materialized.idempotency_index,
-                        &mut materialized.edge_metadata,
-                    );
-                    Ok(())
-                })
-                .await?;
-
-            if base_lsn > last_replayed_lsn {
-                return Err(RepoError::SnapshotNotFound(format!("wal-lsn-{base_lsn}")));
-            }
-        }
-
-        let mut snapshot_catalog = SnapshotCatalog::open(snapshot_catalog_path(&wal_path)).await?;
-        let durable_lsn = {
-            let wal_lock = wal.lock().await;
-            wal_lock.durable_lsn()
-        };
-        snapshot_catalog.truncate_after_lsn(durable_lsn).await?;
-        snapshot_catalog
-            .record_snapshot(durable_lsn, current_unix_timestamp_ms())
-            .await?;
-
-        let storage_capabilities = storage_profile.resolve_capabilities();
-
-        Ok(Self {
-            wal,
-            tx_lock,
-            nodes: Arc::new(RwLock::new(materialized.nodes)),
-            hyper_index: Arc::new(RwLock::new(materialized.hyper_index)),
-            idempotency_index: Arc::new(RwLock::new(materialized.idempotency_index)),
-            edge_metadata: Arc::new(RwLock::new(materialized.edge_metadata)),
-            snapshot_manager,
-            snapshot_catalog: Arc::new(Mutex::new(snapshot_catalog)),
-            session_manager: Arc::new(SessionManager::new(DEFAULT_SESSION_TTL)),
-            storage_profile,
-            storage_capabilities,
-        })
-    }
-
-    pub fn storage_profile(&self) -> &StorageProfile {
-        &self.storage_profile
-    }
-
-    pub fn storage_capabilities(&self) -> &StorageCapabilities {
-        &self.storage_capabilities
-    }
-
-    pub async fn put_node(&self, node: Node) -> Result<(), RepoError> {
-        self.apply_index_transaction(vec![IndexMutation::PutNode(node)])
-            .await
-    }
-
-    pub async fn put_edge(&self, edge: Edge) -> Result<(), RepoError> {
-        self.apply_index_transaction(vec![IndexMutation::PutEdge(edge)])
-            .await
-    }
-
-    pub async fn get_node(&self, id: u64) -> Result<Node, RepoError> {
-        let nodes = self.nodes.read().await;
-        nodes.get(&id).cloned().ok_or(RepoError::NotFound)
-    }
-
-    pub async fn list_node_ids(&self) -> Vec<u64> {
-        let nodes = self.nodes.read().await;
-        let mut out: Vec<u64> = nodes.keys().copied().collect();
-        out.sort_unstable();
-        out
-    }
-
-    pub async fn get_nodes_by_ids(&self, ids: &[u64]) -> Vec<Node> {
-        let nodes = self.nodes.read().await;
-        let mut out: Vec<Node> = ids.iter().filter_map(|id| nodes.get(id).cloned()).collect();
-        out.sort_by_key(|node| node.id);
-        out
-    }
-
-    pub async fn embedding_dimension(&self) -> Option<usize> {
-        let nodes = self.nodes.read().await;
-        nodes
-            .values()
-            .find_map(|node| (!node.embedding.is_empty()).then_some(node.embedding.len()))
-    }
-
-    pub async fn graph_index(&self) -> AdjacencyGraph {
-        let index = self.hyper_index.read().await;
-        index.graph_index.clone()
-    }
-
-    pub async fn delete_node(&self, id: u64) -> Result<(), RepoError> {
-        self.apply_index_transaction(vec![IndexMutation::DeleteNode(id)])
-            .await
-    }
-
-    pub async fn get_node_with_session(
-        &self,
-        id: u64,
-        session_id: Option<&str>,
-    ) -> Result<Node, RepoError> {
-        if let Some(sid) = session_id {
-            if let Some(session) = self.session_manager.get(sid) {
-                if let Some(node) = session.nodes.get(&id) {
-                    return Ok(node.clone());
-                }
-            }
-        }
-        self.get_node(id).await
-    }
-
-    pub fn get_session_with_owner(
-        &self,
-        session_id: &str,
-        owner: Option<&SessionOwner>,
-    ) -> Result<Option<SessionGraph>, RepoError> {
-        let Some(session) = self.session_manager.get(session_id) else {
-            return Ok(None);
-        };
-
-        if let Some(owner) = owner {
-            match session.owner.as_ref() {
-                Some(existing) if existing == owner => {}
-                _ => {
-                    return Err(RepoError::SessionAccessDenied(session_id.to_string()));
-                }
-            }
-        }
-
-        Ok(Some(session.clone()))
-    }
-
-    pub async fn search_vector_with_session_graph(
-        &self,
-        query: &[f32],
-        k: usize,
-        session: Option<&SessionGraph>,
-    ) -> Vec<(u64, f32)> {
-        let mut results = {
-            let index = self.hyper_index.read().await;
-            index.search_vector(query, k)
-        };
-
-        if let Some(session) = session {
-            use alayasiki_core::embedding::cosine_similarity;
-            let mut session_results: Vec<(u64, f32)> = session
-                .nodes
-                .values()
-                .filter_map(|node| {
-                    cosine_similarity(query, &node.embedding).map(|sim| (node.id, sim))
-                })
-                .collect();
-
-            results.append(&mut session_results);
-            results.sort_by(|a, b| {
-                a.0.cmp(&b.0)
-                    .then_with(|| b.1.partial_cmp(&a.1).unwrap_or(std::cmp::Ordering::Equal))
-            });
-            results.dedup_by_key(|(id, _)| *id);
-            results.sort_by(|a, b| b.1.partial_cmp(&a.1).unwrap_or(std::cmp::Ordering::Equal));
-            results.truncate(k);
-        }
-        results
-    }
-
-    pub async fn search_vector_with_session(
-        &self,
-        query: &[f32],
-        k: usize,
-        session_id: Option<&str>,
-    ) -> Vec<(u64, f32)> {
-        let session = session_id.and_then(|sid| self.session_manager.get(sid).map(|s| s.clone()));
-        self.search_vector_with_session_graph(query, k, session.as_ref())
-            .await
-    }
-
-    pub async fn neighbors_with_session_graph(
-        &self,
-        node_id: u64,
-        session: Option<&SessionGraph>,
-    ) -> Vec<(u64, String, f32)> {
-        let mut results: Vec<(u64, String, f32)> = {
-            let index = self.hyper_index.read().await;
-            index
-                .graph_index
-                .neighbors(node_id)
-                .into_iter()
-                .cloned()
-                .collect()
-        };
-        if let Some(session) = session {
-            for edge in &session.edges {
-                if edge.source == node_id {
-                    results.push((edge.target, edge.relation.clone(), edge.weight));
-                }
-            }
-        }
-        results
-    }
-
-    pub async fn neighbors_with_session(
-        &self,
-        node_id: u64,
-        session_id: Option<&str>,
-    ) -> Vec<(u64, String, f32)> {
-        let session = session_id.and_then(|sid| self.session_manager.get(sid).map(|s| s.clone()));
-        self.neighbors_with_session_graph(node_id, session.as_ref())
-            .await
-    }
-
-    pub fn ingest_to_session(&self, session_id: &str, node: Node) {
-        let mut session = self.session_manager.get_or_create(session_id);
-        session.insert_node(node);
-    }
-
-    pub fn ingest_to_session_with_owner(
-        &self,
-        session_id: &str,
-        owner: &SessionOwner,
-        node: Node,
-    ) -> Result<(), RepoError> {
-        let mut session = self.session_manager.get_or_create(session_id);
-        match session.owner.as_ref() {
-            Some(existing) if existing == owner => {}
-            Some(_) => {
-                return Err(RepoError::SessionAccessDenied(session_id.to_string()));
-            }
-            None => {
-                session.owner = Some(owner.clone());
-            }
-        }
-        session.insert_node(node);
-        Ok(())
-    }
-
-    pub fn insert_edge_to_session(&self, session_id: &str, edge: Edge) {
-        let mut session = self.session_manager.get_or_create(session_id);
-        session.insert_edge(edge);
-    }
-
-    pub fn insert_edge_to_session_with_owner(
-        &self,
-        session_id: &str,
-        owner: &SessionOwner,
-        edge: Edge,
-    ) -> Result<(), RepoError> {
-        let mut session = self.session_manager.get_or_create(session_id);
-        match session.owner.as_ref() {
-            Some(existing) if existing == owner => {}
-            Some(_) => {
-                return Err(RepoError::SessionAccessDenied(session_id.to_string()));
-            }
-            None => {
-                session.owner = Some(owner.clone());
-            }
-        }
-        session.insert_edge(edge);
-        Ok(())
-    }
-
-    /// Promote all nodes and edges in a session to persistent storage.
-    /// This is an atomic operation within a single WAL transaction.
-    pub async fn promote_session_to_persistent(&self, session_id: &str) -> Result<(), RepoError> {
-        let session = self
-            .session_manager
-            .take(session_id)
-            .ok_or(RepoError::NotFound)?;
-        let mut session_to_restore = Some(session.clone());
-
-        let nodes = session.nodes.values().cloned().collect::<Vec<_>>();
-        let edges = session.edges.clone();
-
-        let mut mutations = Vec::with_capacity(nodes.len() + edges.len());
-        for node in nodes {
-            mutations.push(IndexMutation::PutNode(node));
-        }
-        for edge in edges {
-            mutations.push(IndexMutation::PutEdge(edge));
-        }
-
-        if let Err(err) = self.apply_index_transaction(mutations).await {
-            if let Some(session) = session_to_restore.take() {
-                self.session_manager.restore(session);
-            }
-            return Err(err);
-        }
-
-        Ok(())
-    }
-
-    /// Apply index updates atomically within one transaction boundary.
-    /// If validation fails, nothing is written to WAL or in-memory indexes.
-    pub async fn apply_index_transaction(
-        &self,
-        mutations: Vec<IndexMutation>,
-    ) -> Result<(), RepoError> {
-        if mutations.is_empty() {
-            return Ok(());
-        }
-
-        // Serialize transaction validation and apply to avoid TOCTOU between concurrent writers.
-        let _tx_guard = self.tx_lock.lock().await;
-
-        self.validate_index_transaction(&mutations).await?;
-
-        let tx_operations = mutations_to_tx_operations(&mutations);
-        let tx_entry = WalEntry::Transaction(tx_operations);
-        let tx_bytes = serialize_wal_entry(&tx_entry)?;
-
-        // Append the full transaction before applying the in-memory mutation set.
-        // Call Repository::flush() before shutdown when using buffered WAL policies.
-        let durable_lsn = {
-            let mut wal = self.wal.lock().await;
-            wal.append(&tx_bytes).await?;
-            wal.durable_lsn()
-        };
-        self.record_durable_snapshot(durable_lsn).await?;
-
-        // Apply in-memory updates under write locks so readers don't observe partial state.
-        let mut nodes = self.nodes.write().await;
-        let mut index = self.hyper_index.write().await;
-        let mut edge_meta = self.edge_metadata.write().await;
-
-        for mutation in mutations {
-            match mutation {
-                IndexMutation::PutNode(node) => {
-                    let id = node.id;
-                    let embedding = node.embedding.clone();
-                    nodes.insert(id, node);
-                    index.insert_node(id, embedding);
-                }
-                IndexMutation::PutEdge(edge) => {
-                    let key = (edge.source, edge.target, edge.relation.clone());
-                    // Always update: replace with new metadata or clear stale provenance
-                    if edge.metadata.is_empty() {
-                        edge_meta.remove(&key);
-                    } else {
-                        edge_meta.insert(key, edge.metadata.clone());
-                    }
-                    index.upsert_edge(edge.source, edge.target, &edge.relation, edge.weight);
-                }
-                IndexMutation::DeleteNode(id) => {
-                    nodes.remove(&id);
-                    index.remove_node(id);
-                    // Remove edge metadata for edges involving the deleted node
-                    edge_meta.retain(|(src, tgt, _), _| *src != id && *tgt != id);
-                }
-            }
-        }
-
-        Ok(())
-    }
-
-    /// Persist a batch of ingested nodes and their idempotency keys in one WAL transaction.
-    pub async fn persist_ingest_batch(
-        &self,
-        nodes_to_put: Vec<Node>,
-        idempotency_records: Vec<(String, Vec<u64>)>,
-    ) -> Result<(), RepoError> {
-        if nodes_to_put.is_empty() && idempotency_records.is_empty() {
-            return Ok(());
-        }
-
-        let _tx_guard = self.tx_lock.lock().await;
-
-        let node_mutations: Vec<IndexMutation> = nodes_to_put
-            .iter()
-            .cloned()
-            .map(IndexMutation::PutNode)
-            .collect();
-        self.validate_index_transaction(&node_mutations).await?;
-
-        let mut idempotency_index = self.idempotency_index.write().await;
-        let new_idempotency_records: Vec<(String, Vec<u64>)> = idempotency_records
-            .into_iter()
-            .filter(|(key, _)| !idempotency_index.contains_key(key))
-            .collect();
-
-        let mut tx_operations = mutations_to_tx_operations(&node_mutations);
-        tx_operations.extend(new_idempotency_records.iter().map(|(key, node_ids)| {
-            TxOperation::RecordIdempotency {
-                key: key.clone(),
-                node_ids: node_ids.clone(),
-            }
-        }));
-
-        if tx_operations.is_empty() {
-            return Ok(());
-        }
-
-        let tx_entry = WalEntry::Transaction(tx_operations.clone());
-        let tx_bytes = serialize_wal_entry(&tx_entry)?;
-
-        let durable_lsn = {
-            let mut wal = self.wal.lock().await;
-            wal.append(&tx_bytes).await?;
-            wal.durable_lsn()
-        };
-        self.record_durable_snapshot(durable_lsn).await?;
-
-        let mut nodes = self.nodes.write().await;
-        let mut index = self.hyper_index.write().await;
-        let mut edge_meta = self.edge_metadata.write().await;
-
-        for operation in &tx_operations {
-            apply_tx_operation(
-                operation,
-                &mut nodes,
-                &mut index,
-                &mut idempotency_index,
-                &mut edge_meta,
-            );
-        }
-
-        Ok(())
-    }
-
-    pub async fn check_idempotency(&self, key: &str) -> Option<Vec<u64>> {
-        let index = self.idempotency_index.read().await;
-        index.get(key).cloned()
-    }
-
-    /// Force pending WAL entries to durable storage.
-    ///
-    /// Call this before graceful shutdown when using buffered flush policies.
-    pub async fn flush(&self) -> Result<(), RepoError> {
-        let _tx_guard = self.tx_lock.lock().await;
-        let durable_lsn = {
-            let mut wal = self.wal.lock().await;
-            wal.flush().await?;
-            wal.durable_lsn()
-        };
-        self.record_durable_snapshot(durable_lsn).await?;
-        Ok(())
-    }
-
-    /// Get metadata for an edge identified by (source, target, relation).
-    pub async fn get_edge_metadata(
-        &self,
-        source: u64,
-        target: u64,
-        relation: &str,
-    ) -> HashMap<String, String> {
-        let edge_meta = self.edge_metadata.read().await;
-        edge_meta
-            .get(&(source, target, relation.to_string()))
-            .cloned()
-            .unwrap_or_default()
-    }
-
-    /// Bulk-get metadata for multiple edges in a single lock acquisition.
-    /// Returns a map from (source, target, relation) to metadata.
-    pub async fn get_edge_metadata_bulk(
-        &self,
-        keys: &[(u64, u64, String)],
-    ) -> HashMap<EdgeMetaKey, HashMap<String, String>> {
-        let edge_meta = self.edge_metadata.read().await;
-        keys.iter()
-            .filter_map(|key| edge_meta.get(key).map(|meta| (key.clone(), meta.clone())))
-            .collect()
-    }
-
-    /// Return the latest durable WAL snapshot id.
-    pub async fn current_snapshot_id(&self) -> String {
-        let wal = self.wal.lock().await;
-        format!("wal-lsn-{}", wal.durable_lsn())
-    }
-
-    pub async fn resolve_snapshot_id_at_or_before(
-        &self,
-        as_of_unix_ms: i64,
-    ) -> Result<String, RepoError> {
-        let catalog = self.snapshot_catalog.lock().await;
-        catalog
-            .resolve_as_of(as_of_unix_ms)
-            .map(|entry| entry.snapshot_id.clone())
-            .ok_or_else(|| RepoError::SnapshotNotFound(format!("as-of-{as_of_unix_ms}")))
-    }
-
-    pub async fn snapshot_catalog_entries(&self) -> Vec<SnapshotCatalogEntry> {
-        let catalog = self.snapshot_catalog.lock().await;
-        catalog.entries().to_vec()
-    }
-
-    async fn record_durable_snapshot(&self, durable_lsn: u64) -> Result<(), RepoError> {
-        let mut catalog = self.snapshot_catalog.lock().await;
-        catalog
-            .record_snapshot(durable_lsn, current_unix_timestamp_ms())
-            .await?;
-        Ok(())
-    }
-
-    /// Create a durable backup snapshot file at the current WAL LSN.
-    pub async fn create_backup_snapshot(&self) -> Result<String, RepoError> {
-        let snapshot_manager = self
-            .snapshot_manager
-            .as_ref()
-            .ok_or(RepoError::SnapshotNotConfigured)?;
-
-        let snapshot = {
-            let _tx_guard = self.tx_lock.lock().await;
-
-            let lsn = {
-                let mut wal = self.wal.lock().await;
-                wal.flush().await?;
-                wal.durable_lsn()
-            };
-            self.record_durable_snapshot(lsn).await?;
-
-            let mut nodes: Vec<Node> = self.nodes.read().await.values().cloned().collect();
-            nodes.sort_by_key(|node| node.id);
-
-            let edges = {
-                let index = self.hyper_index.read().await;
-                collect_backup_edges(&index)
-            };
-
-            let mut idempotency: Vec<BackupIdempotencyRecord> = self
-                .idempotency_index
-                .read()
-                .await
-                .iter()
-                .map(|(key, node_ids)| BackupIdempotencyRecord {
-                    key: key.clone(),
-                    node_ids: node_ids.clone(),
-                })
-                .collect();
-            idempotency.sort_by(|a, b| a.key.cmp(&b.key));
-
-            let mut edge_metadata: Vec<BackupEdgeMetadataRecord> = self
-                .edge_metadata
-                .read()
-                .await
-                .iter()
-                .map(
-                    |((source, target, relation), metadata)| BackupEdgeMetadataRecord {
-                        source: *source,
-                        target: *target,
-                        relation: relation.clone(),
-                        metadata: metadata.clone(),
-                    },
-                )
-                .collect();
-            edge_metadata.sort_by(|a, b| {
-                a.source
-                    .cmp(&b.source)
-                    .then(a.target.cmp(&b.target))
-                    .then(a.relation.cmp(&b.relation))
-            });
-
-            RepositoryBackupSnapshot {
-                lsn,
-                nodes,
-                edges,
-                idempotency,
-                edge_metadata,
-            }
-        };
-
-        let encoded = serialize_backup_snapshot(&snapshot)?;
-        snapshot_manager
-            .create_snapshot(snapshot.lsn, &encoded)
-            .await?;
-
-        Ok(format!("wal-lsn-{}", snapshot.lsn))
-    }
-
-    /// Rebuild in-memory state from the latest backup snapshot plus WAL delta replay.
-    pub async fn restore_from_latest_backup(&self) -> Result<String, RepoError> {
-        if self.snapshot_manager.is_none() {
-            return Err(RepoError::SnapshotNotConfigured);
-        }
-
-        let _tx_guard = self.tx_lock.lock().await;
-        let target_lsn = {
-            let wal = self.wal.lock().await;
-            wal.current_lsn()
-        };
-
-        let (mut materialized, base_lsn) = load_materialized_state_from_backup(
-            self.snapshot_manager.as_ref(),
-            Some(target_lsn),
-            self.storage_profile.clone(),
-        )
-        .await?;
-
-        {
-            let mut wal = self.wal.lock().await;
-            wal.replay(|lsn, data| {
-                if lsn <= base_lsn || lsn > target_lsn {
-                    return Ok(());
-                }
-
-                let archived = rkyv::check_archived_root::<WalEntry>(&data[..])
-                    .map_err(|_| WalError::CorruptEntry)?;
-                let entry: WalEntry = archived.deserialize(&mut rkyv::Infallible).unwrap();
-                apply_replayed_entry(
-                    &entry,
-                    &mut materialized.nodes,
-                    &mut materialized.hyper_index,
-                    &mut materialized.idempotency_index,
-                    &mut materialized.edge_metadata,
-                );
-                Ok(())
-            })
-            .await?;
-        }
-
-        *self.nodes.write().await = materialized.nodes;
-        *self.hyper_index.write().await = materialized.hyper_index;
-        *self.idempotency_index.write().await = materialized.idempotency_index;
-        *self.edge_metadata.write().await = materialized.edge_metadata;
-
-        Ok(format!("wal-lsn-{target_lsn}"))
-    }
-
-    /// Materialize an immutable read view at the specified snapshot.
-    /// Supported format: `wal-lsn-<number>`.
-    pub async fn load_snapshot_view(&self, snapshot_id: &str) -> Result<SnapshotView, RepoError> {
-        let target_lsn = parse_wal_snapshot_lsn(snapshot_id)
-            .ok_or_else(|| RepoError::InvalidSnapshotId(snapshot_id.to_string()))?;
-
-        let current_lsn = {
-            let wal = self.wal.lock().await;
-            wal.current_lsn()
-        };
-        if target_lsn > current_lsn {
-            return Err(RepoError::SnapshotNotFound(snapshot_id.to_string()));
-        }
-
-        let (mut materialized, base_lsn) = load_materialized_state_from_backup(
-            self.snapshot_manager.as_ref(),
-            Some(target_lsn),
-            self.storage_profile.clone(),
-        )
-        .await?;
-
-        let mut wal = self.wal.lock().await;
-        wal.replay(|lsn, data| {
-            if lsn <= base_lsn || lsn > target_lsn {
-                return Ok(());
-            }
-
-            let archived = rkyv::check_archived_root::<WalEntry>(&data[..])
-                .map_err(|_| WalError::CorruptEntry)?;
-            let entry: WalEntry = archived.deserialize(&mut rkyv::Infallible).unwrap();
-            apply_replayed_entry(
-                &entry,
-                &mut materialized.nodes,
-                &mut materialized.hyper_index,
-                &mut materialized.idempotency_index,
-                &mut materialized.edge_metadata,
-            );
-            Ok(())
-        })
-        .await?;
-
-        Ok(SnapshotView {
-            snapshot_id: snapshot_id.to_string(),
-            nodes: materialized.nodes,
-            hyper_index: materialized.hyper_index,
-            edge_metadata: materialized.edge_metadata,
-        })
-    }
-
-    pub async fn record_idempotency(&self, key: &str, node_ids: Vec<u64>) -> Result<(), RepoError> {
-        // Optimization: Lock once and check.
-        // Review feedback suggests removing double-check pattern if racy or checking under write lock.
-        // We will just acquire write lock immediately to be atomic and safe.
-        // Read check optimization is good for read-heavy, but here we expect write if we got this far.
-
-        {
-            let mut index = self.idempotency_index.write().await;
-            if index.contains_key(key) {
-                return Ok(());
-            }
-
-            // Create WalEntry inside write lock to ensure consistency (though WAL append is async).
-            // Actually, we should serialize first to avoid holding lock during serialization?
-            // But we can't check-then-serialize-then-lock-then-write without race.
-            // So we hold lock.
-
-            let entry = WalEntry::IdempotencyKey {
-                key: key.to_string(),
-                node_ids: node_ids.clone(),
-            };
-            // Increase buffer size for large ID lists
-            let mut serializer = AllocSerializer::<4096>::default();
-            serializer
-                .serialize_value(&entry)
-                .map_err(|_| RepoError::Serialization)?;
-            let bytes = serializer.into_serializer().into_inner();
-
-            let durable_lsn = {
-                let mut wal = self.wal.lock().await;
-                wal.append(&bytes).await?;
-                wal.durable_lsn()
-            };
-            self.record_durable_snapshot(durable_lsn).await?;
-
-            index.insert(key.to_string(), node_ids);
-        }
-
-        Ok(())
-    }
-
-    async fn validate_index_transaction(
-        &self,
-        mutations: &[IndexMutation],
-    ) -> Result<(), RepoError> {
-        let nodes = self.nodes.read().await;
-        let mut visible_nodes: HashSet<u64> = nodes.keys().copied().collect();
-
-        for mutation in mutations {
-            match mutation {
-                IndexMutation::PutNode(node) => {
-                    visible_nodes.insert(node.id);
-                }
-                IndexMutation::PutEdge(edge) => {
-                    if !visible_nodes.contains(&edge.source) {
-                        return Err(RepoError::InvalidTransaction(format!(
-                            "edge source {} does not exist",
-                            edge.source
-                        )));
-                    }
-                    if !visible_nodes.contains(&edge.target) {
-                        return Err(RepoError::InvalidTransaction(format!(
-                            "edge target {} does not exist",
-                            edge.target
-                        )));
-                    }
-                }
-                IndexMutation::DeleteNode(id) => {
-                    if !visible_nodes.remove(id) {
-                        return Err(RepoError::NotFound);
-                    }
-                }
-            }
-        }
-
-        Ok(())
-    }
-}
-
-fn mutations_to_tx_operations(mutations: &[IndexMutation]) -> Vec<TxOperation> {
-    mutations
-        .iter()
-        .map(|mutation| match mutation {
-            IndexMutation::PutNode(node) => TxOperation::Put(node.clone()),
-            IndexMutation::PutEdge(edge) => TxOperation::PutEdge(edge.clone()),
-            IndexMutation::DeleteNode(id) => TxOperation::Delete(*id),
-        })
-        .collect()
-}
-
-fn serialize_wal_entry(entry: &WalEntry) -> Result<Vec<u8>, RepoError> {
-    let mut serializer = AllocSerializer::<4096>::default();
-    serializer
-        .serialize_value(entry)
-        .map_err(|_| RepoError::Serialization)?;
-    Ok(serializer.into_serializer().into_inner().to_vec())
-}
-
-fn serialize_backup_snapshot(snapshot: &RepositoryBackupSnapshot) -> Result<Vec<u8>, RepoError> {
-    let mut serializer = AllocSerializer::<4096>::default();
-    serializer
-        .serialize_value(snapshot)
-        .map_err(|_| RepoError::Serialization)?;
-    Ok(serializer.into_serializer().into_inner().to_vec())
-}
-
-async fn deserialize_backup_snapshot(path: &Path) -> Result<RepositoryBackupSnapshot, RepoError> {
-    let bytes = fs::read(path)
-        .await
-        .map_err(|err| RepoError::Snapshot(SnapshotError::Io(err)))?;
-    let archived = rkyv::check_archived_root::<RepositoryBackupSnapshot>(&bytes[..])
-        .map_err(|_| RepoError::Deserialization)?;
-    archived
-        .deserialize(&mut rkyv::Infallible)
-        .map_err(|_| RepoError::Deserialization)
-}
-
-async fn load_materialized_state_from_backup(
-    snapshot_manager: Option<&SnapshotManager>,
-    target_lsn: Option<u64>,
-    storage_profile: StorageProfile,
-) -> Result<(MaterializedState, u64), RepoError> {
-    let empty_state = || MaterializedState {
-        nodes: HashMap::new(),
-        hyper_index: HyperIndex::with_storage_profile(storage_profile.clone()),
-        idempotency_index: HashMap::new(),
-        edge_metadata: HashMap::new(),
-    };
-
-    let Some(manager) = snapshot_manager else {
-        return Ok((empty_state(), 0));
-    };
-
-    let selected = match target_lsn {
-        Some(lsn) => manager.latest_snapshot_at_or_before(lsn).await?,
-        None => manager.latest_snapshot().await?,
-    };
-
-    let Some((snapshot_lsn, path)) = selected else {
-        return Ok((empty_state(), 0));
-    };
-
-    let snapshot = deserialize_backup_snapshot(&path).await?;
-    if snapshot.lsn != snapshot_lsn {
-        return Err(RepoError::Deserialization);
-    }
-
-    let mut nodes = HashMap::new();
-    let mut hyper_index = HyperIndex::with_storage_profile(storage_profile);
-    for node in snapshot.nodes {
-        let id = node.id;
-        hyper_index.insert_node(id, node.embedding.clone());
-        nodes.insert(id, node);
-    }
-
-    for edge in snapshot.edges {
-        hyper_index.upsert_edge(edge.source, edge.target, &edge.relation, edge.weight);
-    }
-
-    let mut idempotency_index = HashMap::new();
-    for record in snapshot.idempotency {
-        idempotency_index.insert(record.key, record.node_ids);
-    }
-
-    let mut edge_metadata = HashMap::new();
-    for record in snapshot.edge_metadata {
-        edge_metadata.insert(
-            (record.source, record.target, record.relation),
-            record.metadata,
-        );
-    }
-
-    Ok((
-        MaterializedState {
-            nodes,
-            hyper_index,
-            idempotency_index,
-            edge_metadata,
-        },
-        snapshot_lsn,
-    ))
-}
-
-fn collect_backup_edges(index: &HyperIndex) -> Vec<BackupEdgeRecord> {
-    let mut edges = Vec::new();
-    for source in index.graph_index.node_ids() {
-        for (target, relation, weight) in index.graph_index.neighbors(source) {
-            edges.push(BackupEdgeRecord {
-                source,
-                target: *target,
-                relation: relation.clone(),
-                weight: *weight,
-            });
-        }
-    }
-
-    edges.sort_by(|a, b| {
-        a.source
-            .cmp(&b.source)
-            .then(a.target.cmp(&b.target))
-            .then(a.relation.cmp(&b.relation))
-    });
-    edges
-}
-
-fn apply_replayed_entry(
-    entry: &WalEntry,
-    node_map: &mut HashMap<u64, Node>,
-    h_index: &mut HyperIndex,
-    idem_map: &mut HashMap<String, Vec<u64>>,
-    edge_meta: &mut HashMap<EdgeMetaKey, HashMap<String, String>>,
-) {
-    match entry {
-        WalEntry::Put(node) => {
-            let id = node.id;
-            let embedding = node.embedding.clone();
-            node_map.insert(id, node.clone());
-            h_index.insert_node(id, embedding);
-        }
-        WalEntry::PutEdge(edge) => {
-            let key = (edge.source, edge.target, edge.relation.clone());
-            if edge.metadata.is_empty() {
-                edge_meta.remove(&key);
-            } else {
-                edge_meta.insert(key, edge.metadata.clone());
-            }
-            h_index.upsert_edge(edge.source, edge.target, &edge.relation, edge.weight);
-        }
-        WalEntry::Delete(id) => {
-            node_map.remove(id);
-            h_index.remove_node(*id);
-            edge_meta.retain(|(src, tgt, _), _| *src != *id && *tgt != *id);
-        }
-        WalEntry::IdempotencyKey { key, node_ids } => {
-            record_idempotency_if_absent(idem_map, key, node_ids);
-        }
-        WalEntry::Transaction(operations) => {
-            for operation in operations {
-                apply_tx_operation(operation, node_map, h_index, idem_map, edge_meta);
-            }
-        }
-    }
-}
-
-fn record_idempotency_if_absent(
-    idem_map: &mut HashMap<String, Vec<u64>>,
-    key: &str,
-    node_ids: &[u64],
-) {
-    idem_map
-        .entry(key.to_string())
-        .or_insert_with(|| node_ids.to_vec());
-}
-
-fn apply_tx_operation(
-    operation: &TxOperation,
-    node_map: &mut HashMap<u64, Node>,
-    h_index: &mut HyperIndex,
-    idem_map: &mut HashMap<String, Vec<u64>>,
-    edge_meta: &mut HashMap<EdgeMetaKey, HashMap<String, String>>,
-) {
-    match operation {
-        TxOperation::Put(node) => {
-            let id = node.id;
-            let embedding = node.embedding.clone();
-            node_map.insert(id, node.clone());
-            h_index.insert_node(id, embedding);
-        }
-        TxOperation::PutEdge(edge) => {
-            let key = (edge.source, edge.target, edge.relation.clone());
-            if edge.metadata.is_empty() {
-                edge_meta.remove(&key);
-            } else {
-                edge_meta.insert(key, edge.metadata.clone());
-            }
-            h_index.upsert_edge(edge.source, edge.target, &edge.relation, edge.weight);
-        }
-        TxOperation::Delete(id) => {
-            node_map.remove(id);
-            h_index.remove_node(*id);
-            edge_meta.retain(|(src, tgt, _), _| *src != *id && *tgt != *id);
-        }
-        TxOperation::RecordIdempotency { key, node_ids } => {
-            record_idempotency_if_absent(idem_map, key, node_ids);
-        }
-    }
-}
-
-fn snapshot_catalog_path(wal_path: &Path) -> PathBuf {
-    wal_path.with_extension("snapshot_catalog.rkyv")
-}
-
-fn current_unix_timestamp_ms() -> i64 {
-    std::time::SystemTime::now()
-        .duration_since(std::time::UNIX_EPOCH)
-        .unwrap_or_default()
-        .as_millis() as i64
-}
-
-pub fn parse_wal_snapshot_lsn(snapshot_id: &str) -> Option<u64> {
-    snapshot_id.strip_prefix("wal-lsn-")?.parse::<u64>().ok()
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::wal::WalFlushPolicy;
-    use std::thread;
-    use std::time::Duration;
-    use tempfile::tempdir;
-
-    #[tokio::test]
-    async fn test_repo_put_get() {
-        let dir = tempdir().unwrap();
-        let wal_path = dir.path().join("test.wal");
-
-        let repo = Repository::open(&wal_path).await.unwrap();
-
-        let node = Node::new(1, vec![0.1, 0.2, 0.3], "Test Node".to_string());
-        repo.put_node(node.clone()).await.unwrap();
-
-        let retrieved = repo.get_node(1).await.unwrap();
-        assert_eq!(retrieved, node);
-    }
-
-    #[tokio::test]
-    async fn test_repo_replay_on_restart() {
-        let dir = tempdir().unwrap();
-        let wal_path = dir.path().join("replay.wal");
-
-        // 1. Create and populate
-        {
-            let repo = Repository::open(&wal_path).await.unwrap();
-            repo.put_node(Node::new(1, vec![1.0], "Node 1".to_string()))
-                .await
-                .unwrap();
-            repo.put_node(Node::new(2, vec![2.0], "Node 2".to_string()))
-                .await
-                .unwrap();
-        }
-
-        // 2. Reopen and verify replay
-        {
-            let repo = Repository::open(&wal_path).await.unwrap();
-            assert_eq!(repo.get_node(1).await.unwrap().data, "Node 1");
-            assert_eq!(repo.get_node(2).await.unwrap().data, "Node 2");
-        }
-    }
-
-    #[tokio::test]
-    async fn test_repo_delete_tombstone() {
-        let dir = tempdir().unwrap();
-        let wal_path = dir.path().join("delete.wal");
-
-        // 1. Create, put, delete
-        {
-            let repo = Repository::open(&wal_path).await.unwrap();
-            repo.put_node(Node::new(1, vec![1.0], "Node 1".to_string()))
-                .await
-                .unwrap();
-            repo.delete_node(1).await.unwrap();
-        }
-
-        // 2. Reopen and verify deleted node is gone
-        {
-            let repo = Repository::open(&wal_path).await.unwrap();
-            assert!(repo.get_node(1).await.is_err());
-        }
-    }
-
-    #[tokio::test]
-    async fn test_repo_edge_and_index_restore() {
-        let dir = tempdir().unwrap();
-        let wal_path = dir.path().join("edge_index.wal");
-
-        // 1. Create Data (Node + Edge)
-        {
-            let repo = Repository::open(&wal_path).await.unwrap();
-            repo.put_node(Node::new(1, vec![1.0, 0.0], "N1".to_string()))
-                .await
-                .unwrap();
-            repo.put_node(Node::new(2, vec![0.0, 1.0], "N2".to_string()))
-                .await
-                .unwrap();
-            repo.put_edge(Edge::new(1, 2, "links", 1.0)).await.unwrap();
-
-            // Verify in-memory index
-            let index = repo.hyper_index.read().await;
-            let neighbors = index.expand_graph(1, 1);
-            assert_eq!(neighbors.len(), 1);
-            assert_eq!(neighbors[0].0, 2);
-        }
-
-        // 2. Restart and Verify Restoration
-        {
-            let repo = Repository::open(&wal_path).await.unwrap();
-
-            // Check HyperIndex restored
-            let index = repo.hyper_index.read().await;
-
-            // Validation 1: Vector Index
-            let search_res = index.search_vector(&[1.0, 0.0], 1);
-            assert_eq!(search_res.len(), 1);
-            assert_eq!(search_res[0].0, 1);
-
-            // Validation 2: Graph Index
-            let neighbors = index.expand_graph(1, 1);
-            assert_eq!(neighbors.len(), 1);
-            assert_eq!(neighbors[0].0, 2);
-        }
-    }
-
-    #[tokio::test]
-    async fn test_index_transaction_commits_all_mutations() {
-        let dir = tempdir().unwrap();
-        let wal_path = dir.path().join("txn_commit.wal");
-        let repo = Repository::open(&wal_path).await.unwrap();
-
-        let n1 = Node::new(1, vec![1.0, 0.0], "N1".to_string());
-        let n2 = Node::new(2, vec![0.0, 1.0], "N2".to_string());
-        let e = Edge::new(1, 2, "links", 1.0);
-
-        repo.apply_index_transaction(vec![
-            IndexMutation::PutNode(n1),
-            IndexMutation::PutNode(n2),
-            IndexMutation::PutEdge(e),
-        ])
-        .await
-        .unwrap();
-
-        assert_eq!(repo.get_node(1).await.unwrap().data, "N1");
-        assert_eq!(repo.get_node(2).await.unwrap().data, "N2");
-
-        let index = repo.hyper_index.read().await;
-        let neighbors = index.expand_graph(1, 1);
-        assert_eq!(neighbors.len(), 1);
-        assert_eq!(neighbors[0].0, 2);
-    }
-
-    #[tokio::test]
-    async fn test_index_transaction_rollback_on_validation_error() {
-        let dir = tempdir().unwrap();
-        let wal_path = dir.path().join("txn_rollback.wal");
-        let repo = Repository::open(&wal_path).await.unwrap();
-
-        let n1 = Node::new(1, vec![1.0], "N1".to_string());
-        let invalid_edge = Edge::new(1, 999, "links", 1.0);
-
-        let result = repo
-            .apply_index_transaction(vec![
-                IndexMutation::PutNode(n1),
-                IndexMutation::PutEdge(invalid_edge),
-            ])
-            .await;
-
-        assert!(result.is_err());
-        assert!(
-            repo.get_node(1).await.is_err(),
-            "node should not be partially committed"
-        );
-
-        let reopened = Repository::open(&wal_path).await.unwrap();
-        assert!(
-            reopened.get_node(1).await.is_err(),
-            "node should not be recoverable after failed transaction"
-        );
-    }
-
-    #[tokio::test]
-    async fn test_index_transaction_persists_single_wal_record() {
-        let dir = tempdir().unwrap();
-        let wal_path = dir.path().join("txn_single_record.wal");
-        let repo = Repository::open(&wal_path).await.unwrap();
-
-        repo.apply_index_transaction(vec![
-            IndexMutation::PutNode(Node::new(1, vec![1.0], "N1".to_string())),
-            IndexMutation::PutNode(Node::new(2, vec![2.0], "N2".to_string())),
-            IndexMutation::PutEdge(Edge::new(1, 2, "links", 1.0)),
-        ])
-        .await
-        .unwrap();
-
-        drop(repo);
-
-        let mut wal = Wal::open(&wal_path).await.unwrap();
-        let mut record_count = 0usize;
-        let mut tx_mutation_count = 0usize;
-
-        wal.replay(|_lsn, payload| {
-            record_count += 1;
-            let archived = rkyv::check_archived_root::<WalEntry>(&payload[..])
-                .map_err(|_| WalError::CorruptEntry)?;
-            let entry: WalEntry = archived.deserialize(&mut rkyv::Infallible).unwrap();
-
-            match entry {
-                WalEntry::Transaction(entries) => {
-                    tx_mutation_count = entries.len();
-                }
-                _ => return Err(WalError::CorruptEntry),
-            }
-
-            Ok(())
-        })
-        .await
-        .unwrap();
-
-        assert_eq!(
-            record_count, 1,
-            "transaction should be written as one WAL record"
-        );
-        assert_eq!(tx_mutation_count, 3);
-    }
-
-    #[tokio::test]
-    async fn test_persist_ingest_batch_persists_nodes_and_idempotency_in_single_wal_record() {
-        let dir = tempdir().unwrap();
-        let wal_path = dir.path().join("ingest_batch_single_record.wal");
-        let repo = Repository::open(&wal_path).await.unwrap();
-
-        let node_ids = vec![11_u64, 12_u64];
-        repo.persist_ingest_batch(
-            vec![
-                Node::new(node_ids[0], vec![1.0], "N1".to_string()),
-                Node::new(node_ids[1], vec![2.0], "N2".to_string()),
-            ],
-            vec![
-                ("content-hash".to_string(), node_ids.clone()),
-                ("request-key".to_string(), node_ids.clone()),
-            ],
-        )
-        .await
-        .unwrap();
-
-        assert_eq!(repo.current_snapshot_id().await, "wal-lsn-1");
-        assert_eq!(
-            repo.check_idempotency("request-key").await,
-            Some(node_ids.clone())
-        );
-
-        drop(repo);
-
-        let reopened = Repository::open(&wal_path).await.unwrap();
-        assert_eq!(reopened.list_node_ids().await, node_ids);
-        assert_eq!(
-            reopened.check_idempotency("content-hash").await,
-            Some(vec![11, 12])
-        );
-
-        let mut wal = Wal::open(&wal_path).await.unwrap();
-        let mut record_count = 0usize;
-        let mut tx_mutation_count = 0usize;
-
-        wal.replay(|_lsn, payload| {
-            record_count += 1;
-            let archived = rkyv::check_archived_root::<WalEntry>(&payload[..])
-                .map_err(|_| WalError::CorruptEntry)?;
-            let entry: WalEntry = archived.deserialize(&mut rkyv::Infallible).unwrap();
-
-            match entry {
-                WalEntry::Transaction(entries) => {
-                    tx_mutation_count = entries.len();
-                }
-                _ => return Err(WalError::CorruptEntry),
-            }
-
-            Ok(())
-        })
-        .await
-        .unwrap();
-
-        assert_eq!(record_count, 1);
-        assert_eq!(tx_mutation_count, 4);
-    }
-
-    #[tokio::test]
-    async fn test_persist_ingest_batch_keeps_first_content_hash_mapping() {
-        let dir = tempdir().unwrap();
-        let wal_path = dir.path().join("ingest_batch_first_writer.wal");
-        let repo = Repository::open(&wal_path).await.unwrap();
-
-        repo.persist_ingest_batch(
-            vec![Node::new(11, vec![1.0], "first".to_string())],
-            vec![
-                ("content-hash".to_string(), vec![11]),
-                ("request-a".to_string(), vec![11]),
-            ],
-        )
-        .await
-        .unwrap();
-
-        repo.persist_ingest_batch(
-            vec![
-                Node::new(11, vec![1.0], "first".to_string()),
-                Node::new(12, vec![2.0], "second".to_string()),
-            ],
-            vec![
-                ("content-hash".to_string(), vec![11, 12]),
-                ("request-b".to_string(), vec![11, 12]),
-            ],
-        )
-        .await
-        .unwrap();
-
-        assert_eq!(repo.check_idempotency("content-hash").await, Some(vec![11]));
-        assert_eq!(repo.check_idempotency("request-a").await, Some(vec![11]));
-        assert_eq!(
-            repo.check_idempotency("request-b").await,
-            Some(vec![11, 12])
-        );
-
-        drop(repo);
-
-        let reopened = Repository::open(&wal_path).await.unwrap();
-        assert_eq!(
-            reopened.check_idempotency("content-hash").await,
-            Some(vec![11])
-        );
-        assert_eq!(
-            reopened.check_idempotency("request-b").await,
-            Some(vec![11, 12])
-        );
-    }
-
-    #[tokio::test]
-    async fn test_index_transaction_flush_and_reopen_preserves_seeded_graph() {
-        let dir = tempdir().unwrap();
-        let wal_path = dir.path().join("txn_seed_flush_reopen.wal");
-        let repo = Repository::open_with_options(
-            &wal_path,
-            WalOptions {
-                flush_policy: WalFlushPolicy::Batch { max_entries: 8 },
-                ..WalOptions::default()
-            },
-        )
-        .await
-        .unwrap();
-
-        repo.apply_index_transaction(vec![
-            IndexMutation::PutNode(Node::new(1, vec![1.0], "N1".to_string())),
-            IndexMutation::PutNode(Node::new(2, vec![2.0], "N2".to_string())),
-            IndexMutation::PutNode(Node::new(3, vec![3.0], "N3".to_string())),
-        ])
-        .await
-        .unwrap();
-        repo.apply_index_transaction(vec![
-            IndexMutation::PutEdge(Edge::new(1, 2, "links", 1.0)),
-            IndexMutation::PutEdge(Edge::new(1, 3, "links", 0.5)),
-        ])
-        .await
-        .unwrap();
-        repo.flush().await.unwrap();
-
-        drop(repo);
-
-        let reopened = Repository::open(&wal_path).await.unwrap();
-        assert_eq!(reopened.list_node_ids().await, vec![1, 2, 3]);
-
-        let graph = reopened.graph_index().await;
-        assert_eq!(graph.edge_count(), 2);
-        let neighbors = graph.neighbors(1);
-        assert_eq!(neighbors.len(), 2);
-        assert!(neighbors
-            .iter()
-            .any(|(target, relation, _)| { *target == 2 && relation == "links" }));
-        assert!(neighbors
-            .iter()
-            .any(|(target, relation, _)| { *target == 3 && relation == "links" }));
-    }
-
-    #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-    async fn test_concurrent_delete_and_put_edge_do_not_leave_dangling_edge() {
-        use tokio::sync::Barrier;
-
-        let dir = tempdir().unwrap();
-        let wal_path = dir.path().join("txn_concurrency.wal");
-        let repo = Arc::new(Repository::open(&wal_path).await.unwrap());
-
-        repo.put_node(Node::new(1, vec![1.0], "N1".to_string()))
-            .await
-            .unwrap();
-
-        for _ in 0..64 {
-            if repo.get_node(2).await.is_err() {
-                repo.put_node(Node::new(2, vec![2.0], "N2".to_string()))
-                    .await
-                    .unwrap();
-            }
-
-            let barrier = Arc::new(Barrier::new(3));
-
-            let repo_put = repo.clone();
-            let barrier_put = barrier.clone();
-            let put_task = tokio::spawn(async move {
-                barrier_put.wait().await;
-                let _ = repo_put.put_edge(Edge::new(1, 2, "links", 1.0)).await;
-            });
-
-            let repo_delete = repo.clone();
-            let barrier_delete = barrier.clone();
-            let delete_task = tokio::spawn(async move {
-                barrier_delete.wait().await;
-                let _ = repo_delete.delete_node(2).await;
-            });
-
-            barrier.wait().await;
-            put_task.await.unwrap();
-            delete_task.await.unwrap();
-
-            assert!(
-                repo.get_node(2).await.is_err(),
-                "node 2 should be deleted after concurrent operations"
-            );
-
-            let index = repo.hyper_index.read().await;
-            let neighbors = index.expand_graph(1, 1);
-            assert!(
-                !neighbors.iter().any(|(id, _)| *id == 2),
-                "dangling edge to deleted node must not remain"
-            );
-        }
-    }
-
-    #[tokio::test]
-    async fn test_load_snapshot_view_reconstructs_historical_state() {
-        let dir = tempdir().unwrap();
-        let wal_path = dir.path().join("snapshot_view.wal");
-        let repo = Repository::open(&wal_path).await.unwrap();
-
-        repo.put_node(Node::new(1, vec![1.0], "N1".to_string()))
-            .await
-            .unwrap();
-        let snapshot_lsn_1 = repo.current_snapshot_id().await;
-
-        repo.put_node(Node::new(2, vec![2.0], "N2".to_string()))
-            .await
-            .unwrap();
-        repo.delete_node(1).await.unwrap();
-
-        let view_at_lsn_1 = repo.load_snapshot_view(&snapshot_lsn_1).await.unwrap();
-        assert_eq!(view_at_lsn_1.snapshot_id(), snapshot_lsn_1);
-        assert_eq!(view_at_lsn_1.list_node_ids(), vec![1]);
-        assert_eq!(view_at_lsn_1.get_nodes_by_ids(&[1])[0].data, "N1");
-
-        let view_at_lsn_3 = repo.load_snapshot_view("wal-lsn-3").await.unwrap();
-        assert_eq!(view_at_lsn_3.list_node_ids(), vec![2]);
-    }
-
-    #[tokio::test]
-    async fn test_load_snapshot_view_rejects_missing_or_invalid_snapshot_id() {
-        let dir = tempdir().unwrap();
-        let wal_path = dir.path().join("snapshot_view_errors.wal");
-        let repo = Repository::open(&wal_path).await.unwrap();
-
-        repo.put_node(Node::new(1, vec![1.0], "N1".to_string()))
-            .await
-            .unwrap();
-
-        let missing = repo.load_snapshot_view("wal-lsn-99").await;
-        assert!(matches!(missing, Err(RepoError::SnapshotNotFound(_))));
-
-        let invalid = repo.load_snapshot_view("snap-custom").await;
-        assert!(matches!(invalid, Err(RepoError::InvalidSnapshotId(_))));
-    }
-
-    #[tokio::test]
-    async fn test_current_snapshot_id_tracks_durable_lsn_for_buffered_policies() {
-        let dir = tempdir().unwrap();
-        let wal_path = dir.path().join("durable_snapshot_id_batch.wal");
-        let repo = Repository::open_with_options(
-            &wal_path,
-            WalOptions {
-                flush_policy: WalFlushPolicy::Batch { max_entries: 8 },
-                ..WalOptions::default()
-            },
-        )
-        .await
-        .unwrap();
-
-        repo.put_node(Node::new(1, vec![1.0], "N1".to_string()))
-            .await
-            .unwrap();
-        assert_eq!(repo.current_snapshot_id().await, "wal-lsn-0");
-
-        repo.flush().await.unwrap();
-        assert_eq!(repo.current_snapshot_id().await, "wal-lsn-1");
-
-        drop(repo);
-
-        let reopened = Repository::open(&wal_path).await.unwrap();
-        assert_eq!(reopened.list_node_ids().await, vec![1]);
-    }
-
-    #[tokio::test]
-    async fn test_resolve_snapshot_id_at_or_before_uses_persisted_catalog() {
-        let before_repo = current_unix_timestamp_ms() - 1;
-        let dir = tempdir().unwrap();
-        let wal_path = dir.path().join("snapshot_catalog_resolution.wal");
-        let repo = Repository::open(&wal_path).await.unwrap();
-
-        repo.put_node(Node::new(1, vec![1.0], "N1".to_string()))
-            .await
-            .unwrap();
-        let first_snapshot = repo.current_snapshot_id().await;
-        let after_first = current_unix_timestamp_ms();
-
-        thread::sleep(Duration::from_millis(5));
-
-        repo.put_node(Node::new(2, vec![2.0], "N2".to_string()))
-            .await
-            .unwrap();
-
-        assert!(matches!(
-            repo.resolve_snapshot_id_at_or_before(before_repo).await,
-            Err(RepoError::SnapshotNotFound(_))
-        ));
-        assert_eq!(
-            repo.resolve_snapshot_id_at_or_before(after_first)
-                .await
-                .unwrap(),
-            first_snapshot
-        );
-        assert_eq!(repo.snapshot_catalog_entries().await.len(), 3);
-    }
-
-    #[tokio::test]
-    async fn test_create_backup_snapshot_flushes_pending_wal_before_persisting() {
-        let dir = tempdir().unwrap();
-        let wal_path = dir.path().join("snapshot_flush_batch.wal");
-        let snapshot_dir = dir.path().join("snapshots");
-
-        {
-            let repo = Repository::open_with_cipher_and_snapshots_and_options(
-                &wal_path,
-                Arc::new(NoOpCipher),
-                &snapshot_dir,
-                WalOptions {
-                    flush_policy: WalFlushPolicy::Batch { max_entries: 8 },
-                    ..WalOptions::default()
-                },
-            )
-            .await
-            .unwrap();
-            repo.put_node(Node::new(1, vec![1.0], "N1".to_string()))
-                .await
-                .unwrap();
-            repo.put_node(Node::new(2, vec![2.0], "N2".to_string()))
-                .await
-                .unwrap();
-
-            let snapshot_id = repo.create_backup_snapshot().await.unwrap();
-            assert_eq!(snapshot_id, "wal-lsn-2");
-        }
-
-        let reopened = Repository::open_with_snapshots(&wal_path, &snapshot_dir)
-            .await
-            .unwrap();
-        assert_eq!(reopened.list_node_ids().await, vec![1, 2]);
-    }
-
-    #[tokio::test]
-    async fn test_open_with_snapshots_restores_snapshot_and_wal_delta() {
-        let dir = tempdir().unwrap();
-        let wal_path = dir.path().join("snapshot_restore.wal");
-        let snapshot_dir = dir.path().join("snapshots");
-
-        {
-            let repo = Repository::open_with_snapshots(&wal_path, &snapshot_dir)
-                .await
-                .unwrap();
-            repo.put_node(Node::new(1, vec![1.0], "N1".to_string()))
-                .await
-                .unwrap();
-            repo.put_node(Node::new(2, vec![2.0], "N2".to_string()))
-                .await
-                .unwrap();
-
-            let backup_id = repo.create_backup_snapshot().await.unwrap();
-            assert_eq!(backup_id, "wal-lsn-2");
-
-            repo.put_node(Node::new(3, vec![3.0], "N3".to_string()))
-                .await
-                .unwrap();
-        }
-
-        let reopened = Repository::open_with_snapshots(&wal_path, &snapshot_dir)
-            .await
-            .unwrap();
-        assert_eq!(reopened.list_node_ids().await, vec![1, 2, 3]);
-    }
-
-    #[tokio::test]
-    async fn test_restore_from_latest_backup_rebuilds_in_memory_state() {
-        let dir = tempdir().unwrap();
-        let wal_path = dir.path().join("restore_latest_from_backup.wal");
-        let snapshot_dir = dir.path().join("snapshots");
-
-        let repo = Repository::open_with_snapshots(&wal_path, &snapshot_dir)
-            .await
-            .unwrap();
-        repo.put_node(Node::new(1, vec![1.0], "N1".to_string()))
-            .await
-            .unwrap();
-        repo.put_node(Node::new(2, vec![2.0], "N2".to_string()))
-            .await
-            .unwrap();
-        repo.create_backup_snapshot().await.unwrap();
-        repo.delete_node(2).await.unwrap();
-
-        // Simulate transient in-memory corruption and verify restore recovers from durable state.
-        repo.nodes.write().await.clear();
-        *repo.hyper_index.write().await = HyperIndex::new();
-        repo.idempotency_index.write().await.clear();
-        repo.edge_metadata.write().await.clear();
-
-        assert!(repo.list_node_ids().await.is_empty());
-
-        let restored_snapshot = repo.restore_from_latest_backup().await.unwrap();
-        assert_eq!(restored_snapshot, "wal-lsn-3");
-        assert_eq!(repo.list_node_ids().await, vec![1]);
-    }
-
-    #[tokio::test]
-    async fn test_backup_requires_snapshot_manager_configuration() {
-        let dir = tempdir().unwrap();
-        let wal_path = dir.path().join("no_snapshot_manager.wal");
-        let repo = Repository::open(&wal_path).await.unwrap();
-
-        let create_backup = repo.create_backup_snapshot().await;
-        assert!(matches!(
-            create_backup,
-            Err(RepoError::SnapshotNotConfigured)
-        ));
-
-        let restore = repo.restore_from_latest_backup().await;
-        assert!(matches!(restore, Err(RepoError::SnapshotNotConfigured)));
-    }
-
-    #[tokio::test]
-    async fn test_open_with_snapshots_rejects_snapshot_newer_than_wal() {
-        let dir = tempdir().unwrap();
-        let wal_path = dir.path().join("snapshot_newer_than_wal.wal");
-        let snapshot_dir = dir.path().join("snapshots");
-
-        {
-            let repo = Repository::open_with_snapshots(&wal_path, &snapshot_dir)
-                .await
-                .unwrap();
-            repo.put_node(Node::new(1, vec![1.0], "N1".to_string()))
-                .await
-                .unwrap();
-            repo.put_node(Node::new(2, vec![2.0], "N2".to_string()))
-                .await
-                .unwrap();
-            let snapshot_id = repo.create_backup_snapshot().await.unwrap();
-            assert_eq!(snapshot_id, "wal-lsn-2");
-        }
-
-        tokio::fs::write(&wal_path, &[]).await.unwrap();
-
-        let reopened = Repository::open_with_snapshots(&wal_path, &snapshot_dir).await;
-        assert!(matches!(
-            reopened,
-            Err(RepoError::SnapshotNotFound(ref snapshot_id)) if snapshot_id == "wal-lsn-2"
-        ));
-    }
-
-    #[tokio::test]
-    async fn test_promote_session_restores_data_on_validation_failure() {
-        let dir = tempdir().unwrap();
-        let wal_path = dir.path().join("session_promote_failure_restore.wal");
-        let repo = Repository::open(&wal_path).await.unwrap();
-
-        let session_id = "session-promote-failure";
-        repo.ingest_to_session(
-            session_id,
-            Node::new(1, vec![1.0, 0.0], "session-node".to_string()),
-        );
-        repo.insert_edge_to_session(session_id, Edge::new(1, 2, "missing_target", 1.0));
-
-        let result = repo.promote_session_to_persistent(session_id).await;
-        assert!(matches!(result, Err(RepoError::InvalidTransaction(_))));
-
-        let restored = repo
-            .session_manager
-            .get(session_id)
-            .expect("session should be restored after failed promote");
-        assert_eq!(restored.nodes.len(), 1);
-        assert_eq!(restored.edges.len(), 1);
-    }
-
-    #[tokio::test]
-    async fn test_session_owner_enforced_for_ingest_and_query() {
-        let dir = tempdir().unwrap();
-        let wal_path = dir.path().join("session_owner_enforcement.wal");
-        let repo = Repository::open(&wal_path).await.unwrap();
-
-        let owner_a = SessionOwner::new("tenant-a", "user-a");
-        let owner_b = SessionOwner::new("tenant-a", "user-b");
-        let session_id = "session-owner";
-
-        repo.ingest_to_session_with_owner(
-            session_id,
-            &owner_a,
-            Node::new(1, vec![1.0], "owner-a-node".to_string()),
-        )
-        .unwrap();
-
-        let denied_write = repo.ingest_to_session_with_owner(
-            session_id,
-            &owner_b,
-            Node::new(2, vec![2.0], "owner-b-node".to_string()),
-        );
-        assert!(matches!(
-            denied_write,
-            Err(RepoError::SessionAccessDenied(_))
-        ));
-
-        let denied_read = repo.get_session_with_owner(session_id, Some(&owner_b));
-        assert!(matches!(
-            denied_read,
-            Err(RepoError::SessionAccessDenied(_))
-        ));
-
-        let allowed_read = repo
-            .get_session_with_owner(session_id, Some(&owner_a))
-            .unwrap()
-            .expect("session should exist for owner");
-        assert_eq!(allowed_read.nodes.len(), 1);
-        assert!(allowed_read.nodes.contains_key(&1));
-    }
-}
